% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta}
\alias{beta}
\title{Beta}
\format{
a single numerical value usually around 2
}
\usage{
beta
}
\description{
Beta for adjustment of tie games in the model. Beta is the shape parameter (also known as the Weibull 'slope') for the Weibull distribution, optimized to fit the ties data. Should be around 2
}
\keyword{datasets}
