% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt2.R
\name{btdatamaker}
\alias{btdatamaker}
\title{data<-dataBT2
format: nestedlist, for each season with list:
winner:
.. team (factor: team name)
.. date (date: game date)
.. game.number (integer: game number for the season
.. is.playoffs (binary: is a playoff game)
.. season (factor: season id (e.g. 20172018))
.. last.game (integer: number of points last game (0-2), new teams get 0)
.. last.20 (integer: # of points in last 20 games (0-40), new teams start with 0)
.. at.home (binary: is at home = 1)
.. CF.20 (numeric ev corsi for per game last 20 games average. new teams get average)
.. CA.20 (numeric ev corsi against per game last 20 games average. new teams get average)
.. Cp.20 (numeric: ev corsi for percent last 20 games. new teams get 0.5 (0-1))
.. hits.20 (numeric: number of hits per game last 20 games. new teams get average)
.. sh.20 (numeric: average number of shots per game last 20 games. new teams get average)
.. sa.20 (numeric: average number of shots against per game last 20 games. new teams get average)
.. sh.p20 (numeric: shooting percentage last 20 games average. new teams get average (0-1))
.. sv.p20 (numeric: save percentage last 20 games average. new teams get average (0-1))
.. days.since.last (numeric: days since last game, max 21)
.. days.till.next (numeric: days till next game, max 21)
.. shift.length (numeric: average shift length over last 20 games)
.. pim (numeric: average pim over last 20 games)
.. PDO.20 (numeric, Average PDO for last 20 games)
loser: same factors as winner
predictors:
.. last.year (integer: number of points last year)
.. conference (factor: conference name)
.. division (factor: division name)
.. value (integer: Forbes team value, in million (300-1500) see https://www.forbes.com/nhl-valuations/list/)
.. last.year.playoffs (binary: made playoffs last year)
.. cap.hit (numeric: % of cap space utilization. see https://www.capfriendly.com/)}
\usage{
btdatamaker()
}
\description{
model: seasonBTm <- BTm(player1 = winner, player2 = loser, formula = ~ at.home + date + game.number + is.playoffs + season + last.game + last.10 + CF.20 + CF.p20 + hits.20 + sh.p20 + sv.p20 + b2b.1 + b2b.2 last.year[Team] + conference[Team] + division[Team] + value[Team] + last.year.playoffs[Team] + cap.hit[Team] + (1|Team), id = Team, data = dataBT2$[season])
}
\details{
https://danieljhocking.wordpress.com/2014/12/03/lags-and-moving-means-in-dplyr/
}
