% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-league-standings.R
\name{buildStats}
\alias{buildStats}
\title{Bulid Stats Table}
\usage{
buildStats(scores)
}
\arguments{
\item{scores}{scores to use to build stats table}
}
\value{
a stats table (as tibble)
}
\description{
Bulid Stats Table
}
