% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{buildTeamColours}
\alias{buildTeamColours}
\title{Rebuild \code{teamColours} package data from source CSV}
\usage{
buildTeamColours()
}
\value{
\code{NULL} (invisibly). Writes updated data when \code{usethis} is
available.
}
\description{
Rebuild \code{teamColours} package data from source CSV
}
\keyword{internal}
