% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.R
\name{calculateEloRatings}
\alias{calculateEloRatings}
\title{Calculate ELO for multiple seasons}
\usage{
calculateEloRatings(schedule, ratings_history = NULL, k = 8,
  mean_value = 1505, new_teams = 1300, regress_strength = 3,
  home_adv = 35)
}
\arguments{
\item{schedule}{The schedule with game results, format as a df with Date, HomeTeam, AwayTeam, Result, Diff}

\item{ratings_history}{Any ratings that we know exist.}

\item{k}{The elo k value}

\item{mean_value}{The mean value to regress towards}

\item{new_teams}{Elo strength of new teams}

\item{regress_strength}{regression amount (1/n)}

\item{home_adv}{The home ice advantage}
}
\value{
Elo at each date for each team
}
\description{
Calculate many seasons of elo. Repeatedly calls eloSeason and does mean regression.
}
\keyword{internal}
