% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{cleanModel}
\alias{cleanModel}
\title{Clean a model}
\usage{
cleanModel(cm)
}
\arguments{
\item{cm}{model to clean}
}
\value{
a smaller model, ready for saving
}
\description{
reduce the size of a model for better long-term storage.
}
