% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics-utils.R
\name{colourDelta}
\alias{colourDelta}
\title{Check Colour Delta}
\usage{
colourDelta(hex1, hex2)
}
\arguments{
\item{hex1}{colour one, as hex code #XXXXXX}

\item{hex2}{colour two, as hex code #XXXXXX}
}
\value{
a value from 0 (identical) to 1 (completely opposite, black & white)
}
\description{
Check the similarity of two colours by hex code. 0 = identical, 1 = opposite (black & white)
}
\examples{
# colour similarity between FLA and TBL primary colours
colourDelta("#041E42", "#002868")
}
