% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-league-simulation.R
\name{compile_predictions}
\alias{compile_predictions}
\title{Compile predictions to one object}
\usage{
compile_predictions(dir = getOption("HockeyModel.prediction.path"))
}
\arguments{
\item{dir}{Directory holding the prediction .RDS files.}
}
\value{
a data frame.
}
\description{
compiles predictions from a group of .RDS files to one data.frame
}
