% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{covid_play_in_solver}
\alias{covid_play_in_solver}
\title{Covid Play-in Solver}
\usage{
covid_play_in_solver(
  covidSeries = HockeyModel::covidSeries,
  covidSchedule = HockeyModel::covidSchedule,
  covidScores = HockeyModel::covidScores
)
}
\arguments{
\item{covidSeries}{Defaults to the built-in covid play-inand round-robin series}

\item{covidSchedule}{the schedule of games}

\item{covidScores}{the result of games, manually updated}
}
\value{
a list of data frames
}
\description{
Due to the unique play-in required for 2019-2020 season during the COVID-19 pandemic, this mashes the required odds for play-in teams and round-robin teams of finishing ahead in their series
}
