% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covidSchedule}
\alias{covidSchedule}
\title{Covid Scheduele}
\format{
a data frame
}
\usage{
covidSchedule
}
\description{
Manually created schedule for 2020 COVID Playoff play-in games
}
\keyword{datasets}
