% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covidScores}
\alias{covidScores}
\title{Covid Scores}
\format{
a data frame
}
\usage{
covidScores
}
\description{
Manually updated scores for east and west round robin play-ins
}
\keyword{datasets}
