% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covidSeries}
\alias{covidSeries}
\title{Covid Series}
\format{
A list of data frames
}
\usage{
covidSeries
}
\description{
Manually Updated series for 2020 COVID Playoff play-in games
}
\keyword{datasets}
