% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataBT}
\alias{dataBT}
\title{Bradley-Terry formatted data
Data formatted for Bradley Terry Model with Generalized Davidson terms}
\format{a data frame}
\usage{
dataBT
}
\description{
Bradley-Terry formatted data
Data formatted for Bradley Terry Model with Generalized Davidson terms
}
\keyword{datasets}
