% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DC_UPSILON_NHL}
\alias{DC_UPSILON_NHL}
\title{Default NHL Dixon-Coles sigmoid time-decay midpoint}
\usage{
DC_UPSILON_NHL
}
\description{
\code{upsilon} is the inflection point of the within-season logistic curve (days).
}
\keyword{internal}
