% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DC_UPSILON_PWHL}
\alias{DC_UPSILON_PWHL}
\title{Default PWHL Dixon-Coles sigmoid time-decay midpoint}
\usage{
DC_UPSILON_PWHL
}
\description{
Default PWHL Dixon-Coles sigmoid time-decay midpoint
}
\keyword{internal}
