% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DC_XI_PWHL}
\alias{DC_XI_PWHL}
\title{Default PWHL Dixon-Coles sigmoid time-decay slope}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DC_XI_PWHL
}
\description{
Default PWHL Dixon-Coles sigmoid time-decay slope
}
\keyword{internal}
