% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-dixon-coles-predict.R
\name{dcExpandedOdds}
\alias{dcExpandedOdds}
\title{Expand DC odds into regulation/OT/SO result components}
\usage{
dcExpandedOdds(lambda, mu, params = NULL, maxgoal = 8)
}
\arguments{
\item{lambda}{(\code{double}) Home expected goals.}

\item{mu}{(\code{double}) Away expected goals.}

\item{params}{(\code{list} or \code{NULL}) Dixon-Coles parameter list.}

\item{maxgoal}{(\code{integer(1)}) Maximum goals included per team.}
}
\value{
(\code{numeric}) Probabilities for home win, home OT, home SO, away SO,
away OT, and away win.
}
\description{
Expand DC odds into regulation/OT/SO result components
}
\keyword{internal}
