% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{dcExpectedGoals}
\alias{dcExpectedGoals}
\title{DC Expected Goals}
\usage{
dcExpectedGoals(home, away, params = NULL)
}
\arguments{
\item{home}{The home team name}

\item{away}{The away team name}

\item{params}{The named list containing m, rho, beta, eta, and k. See \link{updateDC} for information on the params list}
}
\value{
a list of $home and $away expected goals
}
\description{
Given a home and away team, provide expected goals for both.
}
\examples{
dcExpectedGoals("Toronto Maple Leafs", "Ottawa Senators")
}
