% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-dixon-coles-predict.R
\name{DCPredict}
\alias{DCPredict}
\title{DC Predict home/draw/away win}
\usage{
DCPredict(
  home,
  away,
  params = NULL,
  maxgoal = 10,
  scores = HockeyModel::scores,
  expected_mean = NULL,
  season_percent = NULL,
  draws = TRUE
)
}
\arguments{
\item{home}{home team}

\item{away}{away team}

\item{params}{The named list containing m, rho, beta, eta, and k. See \code{\link[=updateModel]{updateModel()}} for information on the params list}

\item{maxgoal}{max number of goals per team}

\item{scores}{optional, if not supplying m & rho, scores used to calculate them.}

\item{expected_mean}{the mean lambda & mu, used only for regression}

\item{season_percent}{the percent complete of the season, used for regression}

\item{draws}{Whether draws are allowed. Default True}
}
\value{
a vector of home win, draw, and away win probability, or if draws=False, a vector of home and away win probability
}
\description{
Using Dixon-Coles' technique, predict odds each of home win, draw, and away win.
}
