% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-dixon-coles-predict.R
\name{DCPredictErrorRecover}
\alias{DCPredictErrorRecover}
\title{Recover fallback expected goals when model prediction fails}
\usage{
DCPredictErrorRecover(team, opponent, homeiceadv = FALSE, m = HockeyModel::m)
}
\arguments{
\item{team}{(\code{character(1)}) Team being predicted.}

\item{opponent}{(\code{character(1)}) Opponent team.}

\item{homeiceadv}{(\code{logical(1)}) Whether to include home-ice coefficient.}

\item{m}{(\code{glm}) Fitted Dixon-Coles model.}
}
\value{
(\code{double(1)}) Fallback expected goals estimate.
}
\description{
Recover fallback expected goals when model prediction fails
}
\keyword{internal}
