% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-dixon-coles-workflow.R
\name{dcPredictMultipleDays}
\alias{dcPredictMultipleDays}
\title{DC Predict Multiple Days}
\usage{
dcPredictMultipleDays(
  start = as.Date(getSeasonStartDate()),
  end = Sys.Date(),
  scores = HockeyModel::scores,
  schedule = HockeyModel::schedule,
  filedir = getOption("HockeyModel.prediction.path"),
  nsims = 1e+05,
  cores = NULL,
  likelihood_graphic = TRUE
)
}
\arguments{
\item{start}{First day to predict. Default start of season}

\item{end}{Last day to predict. Default today}

\item{scores}{HockeyModel::scores}

\item{schedule}{HockeyModel::schedule}

\item{filedir}{Where to save prediction files.}

\item{nsims}{number of simulations to run}

\item{cores}{number of cores in parallel to simulate on}

\item{likelihood_graphic}{Whether to call for creation of likelihood graphic. Default True}
}
\value{
true, if successful
}
\description{
For catching up on daily predictions
}
