% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{dcRealSeasonPredict}
\alias{dcRealSeasonPredict}
\title{DC Simulate season with reevaluation every day}
\usage{
dcRealSeasonPredict(nsims = 1e+05, scores = HockeyModel::scores,
  schedule = HockeyModel::schedule, cores = parallel::detectCores() -
  1)
}
\arguments{
\item{nsims}{number of simulations}

\item{scores}{hisotircal scores}

\item{schedule}{schedule this season}

\item{cores}{number of cores to process}
}
\description{
DC Simulate season with reevaluation every day
}
