% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{DCweights_old}
\alias{DCweights_old}
\title{DC Weight}
\usage{
DCweights_old(dates, currentDate = Sys.Date(), xi = 0.00426)
}
\arguments{
\item{dates}{list of dates to calculate weights.}

\item{currentDate}{date from which to calculate weight}

\item{xi}{tuning factor}
}
\value{
list of weights corresponding to dates
}
\description{
DC Weight
}
\keyword{internal}
