% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-dixon-coles-predict.R
\name{dcxG}
\alias{dcxG}
\title{Compute expected goals from a Dixon-Coles probability matrix}
\usage{
dcxG(home, away, params = NULL, maxgoal = 10)
}
\arguments{
\item{home}{(\code{character(1)}) Home team name.}

\item{away}{(\code{character(1)}) Away team name.}

\item{params}{(\code{list} or \code{NULL}) Dixon-Coles parameter list.}

\item{maxgoal}{(\code{integer(1)}) Maximum goals included per team.}
}
\value{
(\code{list}) Home and away expected goals.
}
\description{
Compute expected goals from a Dixon-Coles probability matrix
}
\keyword{internal}
