% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{doFastDC}
\alias{doFastDC}
\title{Fast DC model}
\usage{
doFastDC(m = NULL, scores = scores)
}
\arguments{
\item{m}{result from doFastFit}

\item{scores}{the historical scores to evaluate}
}
\value{
a res object
}
\description{
Fast DC model
}
