% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhl-api-fetch.R
\name{.buildPWHLTeamColours}
\alias{.buildPWHLTeamColours}
\title{Rebuild \code{pwhlTeamColours} package data from source CSV}
\usage{
.buildPWHLTeamColours()
}
\value{
\code{NULL} (invisibly). Writes updated data when \code{usethis} is
available.
}
\description{
Rebuild \code{pwhlTeamColours} package data from source CSV
}
\keyword{internal}
