% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-api-fetch.R
\name{.clean_names}
\alias{.clean_names}
\title{Normalize team name fields}
\usage{
.clean_names(sc)
}
\arguments{
\item{sc}{(\code{vector} or \code{data.frame}) Values containing team names, or a data
frame with team-name columns.}
}
\value{
(\code{vector} or \code{data.frame}) \code{sc} with normalized historical and
accented team names.
}
\description{
Normalize team name fields
}
\keyword{internal}
