% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-dixon-coles-predict.R
\name{.dcLambda}
\alias{.dcLambda}
\title{DC Expected Goals}
\usage{
.dcLambda(home, away, params = NULL)
}
\arguments{
\item{home}{The home team name}

\item{away}{The away team name}

\item{params}{The named list containing m, rho, beta, eta, and k. See \code{\link[=updateModel]{updateModel()}} for information on the params list}
}
\value{
a list of $home and $away Poisson Lambda values -
}
\description{
Given a home and away team, provide lambda values
}
