% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-core.R
\name{.default_nhl_graphics_dir}
\alias{.default_nhl_graphics_dir}
\title{Default NHL graphics directory}
\usage{
.default_nhl_graphics_dir()
}
\value{
(\code{character(1)}) NHL graphics directory derived from package
options.
}
\description{
Default NHL graphics directory
}
\keyword{internal}
