% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-core.R
\name{.default_pwhl_graphics_dir}
\alias{.default_pwhl_graphics_dir}
\title{Default PWHL graphics directory}
\usage{
.default_pwhl_graphics_dir()
}
\value{
(\code{character(1)}) PWHL graphics directory derived from package
options.
}
\description{
Default PWHL graphics directory
}
\keyword{internal}
