% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-team-season-lookup.R
\name{.getNumGames}
\alias{.getNumGames}
\title{Get the regular-season game count for a season}
\usage{
.getNumGames(season = getCurrentSeason8())
}
\arguments{
\item{season}{(\code{character(1)}) Season ID in eight-digit format.}
}
\value{
(\code{numeric(1)}) Number of regular-season games per team.
}
\description{
Get the regular-season game count for a season
}
\keyword{internal}
