% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-team-season-lookup.R
\name{.getShortTeam}
\alias{.getShortTeam}
\title{Convert long team names to short codes}
\usage{
.getShortTeam(teams, teamColours = HockeyModel::teamColours)
}
\arguments{
\item{teams}{(\code{character}) Long-form team names.}

\item{teamColours}{(\code{data.frame}) Team metadata table.}
}
\value{
(\code{character}) Team short code for each input team.
}
\description{
Convert long team names to short codes
}
\keyword{internal}
