% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-team-season-lookup.R
\name{.getTeamConferences}
\alias{.getTeamConferences}
\title{Get conferences for one or more teams}
\usage{
.getTeamConferences(teams, teamColours = HockeyModel::teamColours)
}
\arguments{
\item{teams}{(\code{character}) Team names.}

\item{teamColours}{(\code{data.frame}) Team metadata table.}
}
\value{
(\code{character}) Conference name for each input team.
}
\description{
Get conferences for one or more teams
}
\keyword{internal}
