% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-plots.R
\name{.point_predict_pwhl}
\alias{.point_predict_pwhl}
\title{Build the PWHL points projection plot}
\usage{
.point_predict_pwhl(
  data_dir = getOption("HockeyModel.prediction.path", "./prediction_results")
)
}
\arguments{
\item{data_dir}{(\code{character(1)}) Root prediction directory.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Build the PWHL points projection plot
}
\keyword{internal}
