% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-plots.R
\name{.president_odds_nhl}
\alias{.president_odds_nhl}
\title{Build the NHL first-place plot}
\usage{
.president_odds_nhl(
  data_dir = getOption("HockeyModel.prediction.path", "./prediction_results")
)
}
\arguments{
\item{data_dir}{(\code{character(1)}) Root prediction directory.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Build the NHL first-place plot
}
\keyword{internal}
