% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-dixon-coles-predict.R
\name{.prob_matrix}
\alias{.prob_matrix}
\title{Probability Matrix}
\usage{
.prob_matrix(lambda, mu, params, maxgoal)
}
\arguments{
\item{lambda}{home lambda}

\item{mu}{away mu}

\item{params}{The named list containing m, rho, beta, eta, and k. See \code{\link[=updateModel]{updateModel()}} for information on the params list}

\item{maxgoal}{max goals per game}
}
\value{
a square matrix of maxgoal:maxgoal, with all entries in \verb{[0, 1]} and
summing to 1 (see \link{.validateProbMatrix})
}
\description{
Given a mu, lambda, rho, and theta, generate a probability matrix. Differs from .dcProbMatrix in that no regresson or solving for supplied teams happens
}
