% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-plots.R
\name{.ratings_nhl}
\alias{.ratings_nhl}
\title{Build the NHL ratings plot}
\usage{
.ratings_nhl(m = HockeyModel::m)
}
\arguments{
\item{m}{(\code{any}) NHL model \code{m}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Build the NHL ratings plot
}
\keyword{internal}
