% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-plots.R
\name{.ratings_pwhl}
\alias{.ratings_pwhl}
\title{Build the PWHL ratings plot}
\usage{
.ratings_pwhl(m = HockeyModel::pwhl_m)
}
\arguments{
\item{m}{(\code{any}) PWHL model \code{m}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Build the PWHL ratings plot
}
\keyword{internal}
