% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-dixon-coles-predict.R
\name{.sampleResult}
\alias{.sampleResult}
\title{Sample one or more result labels from expanded win-state probabilities}
\usage{
.sampleResult(hw, hot, hso, aso, aot, aw, size = 1)
}
\arguments{
\item{hw}{(\code{double}) Regulation home-win probability.}

\item{hot}{(\code{double}) Home overtime-win probability.}

\item{hso}{(\code{double}) Home shootout-win probability.}

\item{aso}{(\code{double}) Away shootout-win probability.}

\item{aot}{(\code{double}) Away overtime-win probability.}

\item{aw}{(\code{double}) Regulation away-win probability.}

\item{size}{(\code{integer(1)}) Number of samples to draw.}
}
\value{
(\code{numeric}) Result values encoded like \code{scores$Result}.
}
\description{
Sample one or more result labels from expanded win-state probabilities
}
\keyword{internal}
