% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-plots.R
\name{.today_odds_plot_pwhl}
\alias{.today_odds_plot_pwhl}
\title{Build the PWHL today-odds plot}
\usage{
.today_odds_plot_pwhl(
  date = Sys.Date(),
  params = NULL,
  schedule = HockeyModel::pwhlSchedule,
  scores = HockeyModel::pwhlScores
)
}
\arguments{
\item{date}{(\code{Date}) Target date.}

\item{params}{(\code{list} or \code{NULL}) PWHL model parameters.}

\item{schedule}{(\code{data.frame}) PWHL schedule.}

\item{scores}{(\code{data.frame}) PWHL scores.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object or \code{NULL}.
}
\description{
Build the PWHL today-odds plot
}
\keyword{internal}
