% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-core.R
\name{.update_predictions_pwhl}
\alias{.update_predictions_pwhl}
\title{Update saved PWHL prediction snapshots}
\usage{
.update_predictions_pwhl(
  data_dir = file.path(getOption("HockeyModel.prediction.path", "./prediction_results"),
    "pwhl"),
  scores = HockeyModel::pwhlScores,
  schedule = HockeyModel::pwhlSchedule,
  params = NULL
)
}
\arguments{
\item{data_dir}{(\code{character(1)}) Directory of saved PWHL prediction files.}

\item{scores}{(\code{data.frame}) PWHL scores.}

\item{schedule}{(\code{data.frame}) PWHL schedule.}

\item{params}{(\code{list} or \code{NULL}) PWHL model parameters.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Update saved PWHL prediction snapshots
}
\keyword{internal}
