% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elos}
\alias{elos}
\title{Elo History}
\format{
A sparse dataframe with columns for each historical and active team (WHL and NHL) and each row a date
}
\usage{
elos
}
\description{
Elo rankings for each team through time.
}
\keyword{datasets}
