% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.R
\name{eloSeason}
\alias{eloSeason}
\title{Calculate ELO for one season}
\usage{
eloSeason(schedule, dates, ratings, new_teams, k, home_adv)
}
\arguments{
\item{schedule}{The seasons's schedule (of completed games)}

\item{dates}{The dates of each game}

\item{ratings}{The beginning ratings}

\item{new_teams}{Any new teams}

\item{k}{The ELO k factor}

\item{home_adv}{THe home ice advantage}
}
\value{
Elo at each date for each team
}
\description{
Calculate Elo for one season only. Doesn't include regression to mean
}
\keyword{internal}
