% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eta}
\alias{eta}
\title{Eta}
\format{
a single numerical value usually around 3
}
\usage{
eta
}
\description{
Eta for adjustment of tie games in the model. Eta is the Weibull 'scale' parameter, the distribution is multiplied by the diagonal to get it to better estimate tie game odds. Should be around 3
}
\keyword{datasets}
