% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_predictions}
\alias{example_predictions}
\title{Example Predictions
A example 'predictions' compiled for each team's change in predictions over the 2020-2021 season (Contains predictions from 2021-01-12 to 2021-05-08)}
\format{
a tibble
}
\usage{
example_predictions
}
\description{
Example Predictions
A example 'predictions' compiled for each team's change in predictions over the 2020-2021 season (Contains predictions from 2021-01-12 to 2021-05-08)
}
\keyword{datasets}
