% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_raw_predictions}
\alias{example_raw_predictions}
\title{Example Raw Predictions
A example 'raw predictions' compiled for testing purposes}
\format{
a tibble
}
\usage{
example_raw_predictions
}
\description{
Example Raw Predictions
A example 'raw predictions' compiled for testing purposes
}
\keyword{datasets}
