% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{fastDCPredict}
\alias{fastDCPredict}
\title{DC Predict home/draw/away win}
\usage{
fastDCPredict(home, away, m = NULL, res = NULL, maxgoal = 8,
  scores = scores)
}
\arguments{
\item{home}{home team}

\item{away}{away team}

\item{m}{result from doFastFit}

\item{res}{result from doFastDC}

\item{maxgoal}{max number of goals per team}
}
\value{
a list of home win, draw, and away win probability
}
\description{
Using Dixon Coles technique, predict odds each of home win, draw, and away win.
}
