% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley-terry.R
\name{fitBT}
\alias{fitBT}
\title{Fit BradleyTerry Generalized Davidson Model}
\usage{
fitBT(btdata = prepareDataForBT(), subset_from = "2010-08-01")
}
\arguments{
\item{btdata}{data from prepareDataForBT, or defaults to all data prepared.}

\item{subset_from}{date from which (forward) to run BT model. Full nhl history is at least > 45 GB ram requirement (true value unknown). Runing since 1980-08-01 requires minimum 16GB RAM system RAM, process time = hours. Minimum few years required for estimate of team strength, home advantage, and tie percents}
}
\value{
a gnm fit model
}
\description{
Fit BradleyTerry Generalized Davidson Model
}
