% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley-terry.R
\name{fitBTSimple}
\alias{fitBTSimple}
\title{Fit Simple BradleyTerry Model}
\usage{
fitBTSimple(scores = HockeyModel::scores, subset_from = "2016-08-01")
}
\arguments{
\item{subset_from}{date from which (forward) to run BT model.}

\item{btdata}{data from prepareDataForBT, or defaults to all data prepared.}
}
\value{
a gnm fit model
}
\description{
Fit Simple BradleyTerry Model
}
