% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fittedBT}
\alias{fittedBT}
\title{Generalized Davidson Bradley-Terry model
Results of fitting Bradley Terry Model with Generalized Davidson terms}
\format{a gnm Bradley Terry Davidson model}
\usage{
fittedBT
}
\description{
Generalized Davidson Bradley-Terry model
Results of fitting Bradley Terry Model with Generalized Davidson terms
}
\keyword{datasets}
