% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fittedBTSimple}
\alias{fittedBTSimple}
\title{Simple Bradley-Terry model
Results of fitting Bradley Terry Model}
\format{a gnm Bradley Terry Davidson model}
\usage{
fittedBTSimple
}
\description{
Simple Bradley-Terry model
Results of fitting Bradley Terry Model
}
\keyword{datasets}
