% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{gameIDValidator}
\alias{gameIDValidator}
\title{Validate GameID numbers}
\usage{
gameIDValidator(gameIDs)
}
\arguments{
\item{gameIDs}{a single game ID or vector of game IDs}
}
\value{
TRUE if gameIDs is valid, or FALSE if gameIDs is not (or a vector of length of input of TRUE/FALSE
}
\description{
Validate GameID numbers
}
