% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gamma}
\alias{gamma}
\title{Gamma}
\format{
a single numerical value
}
\usage{
gamma
}
\description{
Gamma for adjustment of tie games in the model. Gamma is a multiplication value to the Poisson multiplier to the diagonal to get it to better estimate tie game odds. should be around 5 or 6
}
\keyword{datasets}
