% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-league-playoffs.R
\name{getCompletedSeries}
\alias{getCompletedSeries}
\title{Build winner/loser summary for completed playoff series}
\usage{
getCompletedSeries(currentSeries)
}
\arguments{
\item{currentSeries}{(\code{data.frame}) Series table from \code{\link[=getAPISeries]{getAPISeries()}}.}
}
\value{
(\code{data.frame}) Series identifier with winner and loser teams.
}
\description{
Build winner/loser summary for completed playoff series
}
\keyword{internal}
