% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getConference}
\alias{getConference}
\title{Get the conference of a team or vector of teams}
\usage{
getConference(team)
}
\arguments{
\item{team}{a single team or vector of teams}
}
\value{
the conference (or vector of conferences) for the team(s)
}
\description{
Get the conference of a team or vector of teams
}
