% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-team-season-lookup.R
\name{getConferences}
\alias{getConferences}
\title{Get conference names from the team lookup table}
\usage{
getConferences(teamColours = HockeyModel::teamColours)
}
\arguments{
\item{teamColours}{(\code{data.frame}) Team metadata table with a \code{Conference}
column.}
}
\value{
(\code{character}) Unique conference names.
}
\description{
Get conference names from the team lookup table
}
\keyword{internal}
