% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-team-season-lookup.R
\name{getCurrentSeason8}
\alias{getCurrentSeason8}
\title{Get the latest NHL season ID}
\usage{
getCurrentSeason8()
}
\value{
(\code{character(1)}) Season ID in eight-digit format such as
\code{"20172018"}.
}
\description{
Get the latest NHL season ID
}
