% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCurrentSeasonStartDate}
\alias{getCurrentSeasonStartDate}
\title{Get Current Season Start Date}
\usage{
getCurrentSeasonStartDate()
}
\value{
current season's nominal start date (Oct 01) as character, e.g. "2019-10-01"
}
\description{
Get Current Season Start Date
}
