% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDivision}
\alias{getDivision}
\title{Get the division of a team or vector of teams}
\usage{
getDivision(team)
}
\arguments{
\item{team}{a single team or vector of teams}
}
\value{
the division (or vector of divisions) for the team(s)
}
\description{
Get the division of a team or vector of teams
}
