% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-team-season-lookup.R
\name{getDivisions}
\alias{getDivisions}
\title{Get division names from the team lookup table}
\usage{
getDivisions(teamColours = HockeyModel::teamColours)
}
\arguments{
\item{teamColours}{(\code{data.frame}) Team metadata table with a \code{Division}
column.}
}
\value{
(\code{character}) Unique division names.
}
\description{
Get division names from the team lookup table
}
\keyword{internal}
