% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-team-season-lookup.R
\name{getLongTeam}
\alias{getLongTeam}
\title{Convert short team codes to long names}
\usage{
getLongTeam(teams, teamColours = HockeyModel::teamColours)
}
\arguments{
\item{teams}{(\code{character}) Team short codes.}

\item{teamColours}{(\code{data.frame}) Team metadata table.}
}
\value{
(\code{character}) Long-form team name for each input code.
}
\description{
Convert short team codes to long names
}
\keyword{internal}
