% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_iterative.R
\name{getNewIterativeRankings}
\alias{getNewIterativeRankings}
\title{Refit iterative rankings for one target model}
\usage{
getNewIterativeRankings(
  target_model = "wl",
  params = HockeyModel::iterativeParameters
)
}
\arguments{
\item{target_model}{(\code{character(1)}) One of \code{"wl"} or \code{"xg"} (case
insensitive variants accepted).}

\item{params}{(\code{list}) Iterative parameter object.}
}
\value{
(\code{data.frame}) Re-estimated team rankings.
}
\description{
Refit iterative rankings for one target model
}
\keyword{internal}
