% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-api-fetch.R
\name{getNHLScores}
\alias{getNHLScores}
\title{Get NHL Scores}
\usage{
getNHLScores(gameIDs = NULL, schedule = HockeyModel::schedule, progress = TRUE)
}
\arguments{
\item{gameIDs}{Game IDs (10 digit number).}

\item{schedule}{optional, provide a schedule if not using the HockeyModel::schedule}

\item{progress}{whether to show a progress bar. Requires the 'progress' package installed}
}
\value{
a data frame with Date, HomeTeam, AwayTeam, GameID, HomeGoals, AwayGoals, OTStatus and GameType
}
\description{
Get the NHL game scores from the NHL API for any game(s) with a final score. Returns data formatted
for future use. Requires gameID(s) be provided.This can be slow if requesting many games due to the API rate limit.
}
