% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_iterative.R
\name{getReplacementIterativeParameters}
\alias{getReplacementIterativeParameters}
\title{Get replacement/new iterative parameters}
\usage{
getReplacementIterativeParameters(
  params = HockeyModel::iterativeParameters,
  save_data = FALSE
)
}
\arguments{
\item{params}{parameters to start with.}

\item{save_data}{whether to save data to the package. Default false}
}
\value{
invisibly, new parameters
}
\description{
Get replacement/new iterative parameters
}
