% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_iterative.R
\name{getReplacementRankings}
\alias{getReplacementRankings}
\title{Get (and optionally save to package) iterative rankings}
\usage{
getReplacementRankings(
  params = HockeyModel::iterativeParameters,
  save_data = FALSE
)
}
\arguments{
\item{params}{input parameters to use for retraining}

\item{save_data}{whether to save the new parameters to package}
}
\value{
invisibly, a new set of rankings
}
\description{
Get (and optionally save to package) iterative rankings
}
