% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-dixon-coles-fit.R
\name{getRho}
\alias{getRho}
\title{Generate a 'rho' factor for low scoring games}
\usage{
getRho(m = HockeyModel::m, scores = HockeyModel::scores)
}
\arguments{
\item{m}{result from getM}

\item{scores}{the historical scores to evaluate}
}
\value{
a numeric value (typically -0.5 to 0)
}
\description{
Generate a 'rho' factor for low scoring games
}
