% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-team-season-lookup.R
\name{getSeasonStartDate}
\alias{getSeasonStartDate}
\title{getSeasonStartDate}
\usage{
getSeasonStartDate(season = NULL)
}
\arguments{
\item{season}{Season (8 character code) or NULL for current/most recent season}
}
\value{
Season start date as date
}
\description{
getSeasonStartDate
}
