% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-league-playoffs.R
\name{getSeriesOdds}
\alias{getSeriesOdds}
\title{Get Series Odds}
\usage{
getSeriesOdds(params = NULL)
}
\arguments{
\item{params}{The named list containing m, rho, beta, eta, and k. See \code{\link[=updateModel]{updateModel()}} for information on the params list}
}
\value{
NULL if no series are currently set but not complete, else a data frame.
}
\description{
Get Series Odds
}
