% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_iterative.R
\name{getTeamRankings}
\alias{getTeamRankings}
\title{Get Team Iterative Rankings}
\usage{
getTeamRankings(team, rankings)
}
\arguments{
\item{team}{Team name}

\item{rankings}{rankings data frame}
}
\value{
attack & defence ratings, plust the rankings data frame again (in case a team is added with attack & defense of 0)
}
\description{
Gets a team's attack and defense ratings for iterative dixon-coles techniques. If a team hasn't been sampled before in this
given rankings data set, attack and defense values of 0 are returned, and the team added to the rankings data frame. The rankings
data frame is always returned as well, it's good tocapture that in case the team is new.
}
