% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{getTheta}
\alias{getTheta}
\title{Get Theta}
\usage{
getTheta(
  m = HockeyModel::m,
  rho = HockeyModel::rho,
  scores = HockeyModel::scores
)
}
\arguments{
\item{m}{HockeyModel::m}

\item{rho}{HockeyMoel::rho}

\item{scores}{HockeyModel::scores}
}
\value{
a theta and gamma value as a list
}
\description{
Theta is the diagonal enhancement for goals, helping to more accurately predict tie games, gamma is the multiplication factor
}
