% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-dixon-coles-fit.R
\name{getWeibullParams}
\alias{getWeibullParams}
\title{Get Weibull Params}
\usage{
getWeibullParams(
  m = HockeyModel::m,
  rho = HockeyModel::rho,
  scores = HockeyModel::scores
)
}
\arguments{
\item{m}{HockeyModel::m}

\item{rho}{HockeyMoel::rho}

\item{scores}{HockeyModel::scores}
}
\value{
a \link{beta} and \link{eta} and \link{k} value as a list
}
\description{
Weibull Params determine is the diagonal enhancement for goals, helping to more accurately predict tie games.
}
