% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gId}
\alias{gId}
\title{Validate one NHL game ID value}
\usage{
gId(gameId)
}
\arguments{
\item{gameId}{(\code{numeric(1)}) Candidate game ID.}
}
\value{
(\code{logical(1)}) \code{TRUE} when the ID matches expected NHL format.
}
\description{
Validate one NHL game ID value
}
\keyword{internal}
