% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics-utils.R
\name{hexToRGB}
\alias{hexToRGB}
\title{Hex to RGB}
\usage{
hexToRGB(hex)
}
\arguments{
\item{hex}{colour as hex code #XXXXXX}
}
\value{
vector of three numbers, R, G, B from 0 to 255
}
\description{
convert Hex colours to RGB colours
}
\examples{
hexToRGB("#FFFFFF")
}
