% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{historicalPoints}
\alias{historicalPoints}
\title{Get Historical Points for all teams listed in a scores frame}
\usage{
historicalPoints(sc)
}
\arguments{
\item{sc}{scores frame}
}
\value{
A tibble with season point totals by team.
}
\description{
Get Historical Points for all teams listed in a scores frame
}
