% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-team-season-lookup.R
\name{inOffSeason}
\alias{inOffSeason}
\title{In Off Season}
\usage{
inOffSeason(date = Sys.Date())
}
\arguments{
\item{date}{Date to check if it's in any off season. Default today}
}
\value{
TRUE if we're in off-season, else FALSE
}
\description{
Determine if a provided date is in the off-season i.e. not in regular season or playoffs
}
