% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-team-season-lookup.R
\name{inRegularSeason}
\alias{inRegularSeason}
\title{In Regular Season}
\usage{
inRegularSeason(date = Sys.Date(), boolean = TRUE)
}
\arguments{
\item{date}{Date to check if it's in a season}

\item{boolean}{Whether to return a result as TRUE/FALSE (True) or to return the seasonID if the date is in a season}
}
\value{
Either TRUE/FALSE or a seasonID/FALSE
}
\description{
In Regular Season
}
