% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_iterative.R
\name{iterateiveDailyUpdate}
\alias{iterateiveDailyUpdate}
\title{Run daily update pipeline for iterative predictions}
\usage{
iterateiveDailyUpdate()
}
\value{
(\code{data.frame} or \code{NULL}) Current-day iterative prediction table.
}
\description{
Run daily update pipeline for iterative predictions
}
\keyword{internal}
