% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_iterative.R
\name{iterative_season_progress}
\alias{iterative_season_progress}
\title{Convert game IDs into coarse season progress values}
\usage{
iterative_season_progress(gameids)
}
\arguments{
\item{gameids}{(\code{character} or \code{numeric}) One or more NHL game IDs.}
}
\value{
(\code{double}) Values in \verb{[0, 1]} describing regular-season progress;
playoff games return \code{1}.
}
\description{
Convert game IDs into coarse season progress values
}
\keyword{internal}
