% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_iterative.R
\name{iterativeGamePredict}
\alias{iterativeGamePredict}
\title{Predict the outcome of one game using iterative Dixon Coles model}
\usage{
iterativeGamePredict(
  home,
  away,
  rankings,
  intercept = 1.71418,
  home_adv = 0.075073,
  probs_style = "h"
)
}
\arguments{
\item{home}{The Home Team}

\item{away}{The Away Team}

\item{rankings}{A rankings dataframe}

\item{intercept}{Modelled 'intercept' value}

\item{home_adv}{Modelled 'home_adv' value}

\item{probs_style}{either \code{'h'} (default) or one of \code{'ha'}, \verb{'hda}, or \code{'hdda'}}
}
\value{
a named list with Probs (a vector), HomexGPred and AwayxGPred (values)
}
\description{
Predict the outcome of one game using iterative Dixon Coles model
}
