% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iterativeParameters}
\alias{iterativeParameters}
\title{Iterative Parameters
Parameters for iterative Dixon-Coles method}
\format{
a list of lists of parameters for win/loss and xG versions
}
\usage{
iterativeParameters
}
\description{
Iterative Parameters
Parameters for iterative Dixon-Coles method
}
\keyword{datasets}
