% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iterativeRankings}
\alias{iterativeRankings}
\title{Iterative Rankings
Current rankings for iterative Dixon-Coles methods}
\format{
a list of data frames
}
\usage{
iterativeRankings
}
\description{
Iterative Rankings
Current rankings for iterative Dixon-Coles methods
}
\keyword{datasets}
