% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{k}
\alias{k}
\title{k}
\format{
a single numerical value usually around 5 or 6
}
\usage{
k
}
\description{
k for adjustment of tie games in the model. k is a multiplication value to the Weibull multiplier to the diagonal to get it to better estimate tie game odds. Should be around 5 or 6
}
\keyword{datasets}
