% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-league-simulation.R
\name{loopless_sim}
\alias{loopless_sim}
\title{'Loopless' simulation}
\usage{
loopless_sim(
  nsims = 1e+05,
  cores = NULL,
  schedule = HockeyModel::schedule,
  scores = HockeyModel::scores,
  params = NULL,
  season_sofar = NULL,
  likelihood_graphic = TRUE,
  odds_table = NULL
)
}
\arguments{
\item{nsims}{number of simulations to run (approximate)}

\item{cores}{number of cores in parallel to process}

\item{schedule}{games to play}

\item{scores}{Season to this point}

\item{params}{The named list containing m, rho, beta, eta, and k. See \code{\link[=updateModel]{updateModel()}} for information on the params list}

\item{season_sofar}{The results of the season to date}

\item{likelihood_graphic}{whether to create a likelihood graphic}

\item{odds_table}{a table of odds for all games in schedule. Null, unless provided. Should be similar to the output of \code{remainderSeasonDC(odds=TRUE)},
which is a data.frame of HomeTeam, AwayTeam, HomeWin, AwayWin, Draw, GameID and Date}
}
\value{
a two member list, of all results and summary results
}
\description{
'Loopless' simulation
}
