% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m}
\alias{m}
\title{Dixon Coles' \code{m}}
\format{
a glm model fit
}
\usage{
m
}
\description{
Model fit for each team's attack and defend strength, as well as home ice advantage.
}
\keyword{datasets}
