% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_cond}
\alias{mutate_cond}
\title{Conditional Mutate}
\usage{
mutate_cond(.data, condition, ..., envir = parent.frame())
}
\arguments{
\item{.data}{Data passed in}

\item{condition}{Condition whether to peform mutate}

\item{...}{mutate to happen}

\item{envir}{environment to cary through.}
}
\description{
Mutate at condition. useful in dplyr pipes.
From StackOverflow https://stackoverflow.com/a/34096575/3933405
}
