% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-api-fetch.R
\name{nhl_boxscore}
\alias{nhl_boxscore}
\title{Fetch an NHL boxscore payload}
\usage{
nhl_boxscore(gid)
}
\arguments{
\item{gid}{(\code{character(1)} or \code{numeric(1)}) NHL game ID.}
}
\value{
(\code{list}) Parsed boxscore response from the NHL API.
}
\description{
Fetch an NHL boxscore payload
}
\keyword{internal}
