% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhl_conferences}
\alias{nhl_conferences}
\title{NHL Conferences}
\format{
A list of vectors:
\describe{
  \item{East}{Teams in the East conference}
  \item{West}{Teams in the West conference}
}
}
\source{
\url{http://www.nhl.com/}
}
\usage{
nhl_conferences
}
\description{
A list of Conferences with teams in each vector.
}
\keyword{datasets}
