% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{normalizeOdds}
\alias{normalizeOdds}
\title{Normalize Odds}
\usage{
normalizeOdds(odds)
}
\arguments{
\item{odds}{a vector of odds to normalize}
}
\value{
odds summing to 1
}
\description{
Normalize Odds
}
