% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_iterative.R
\name{optimizeIterative_WL_Internal}
\alias{optimizeIterative_WL_Internal}
\title{Objective function for iterative win/loss optimization}
\usage{
optimizeIterative_WL_Internal(params)
}
\arguments{
\item{params}{(\code{double}) Candidate parameter vector.}
}
\value{
(\code{double(1)}) Optimization score (lower is better).
}
\description{
Objective function for iterative win/loss optimization
}
\keyword{internal}
