% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{playoffRoundSolver}
\alias{playoffRoundSolver}
\title{Playoff Round Solver}
\usage{
playoffRoundSolver(teamodds)
}
\arguments{
\item{teamodds}{a data frame of each team that could play in a round, with their odds of being in the round as home or away team. Columns \code{Team}, \code{HomeOdds}, \code{AwayOdds}}
}
\value{
a data frame with Team and odds of winning round
}
\description{
Playoff Round Solver
}
