% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-league-playoffs.R
\name{playoffSeriesOdds}
\alias{playoffSeriesOdds}
\title{Statistical Playoff Series Odds Solver}
\usage{
playoffSeriesOdds(
  home_odds,
  away_odds,
  home_win = 0,
  away_win = 0,
  ngames = NULL,
  game_home = NULL,
  predict_games_to_win = FALSE
)
}
\arguments{
\item{home_odds}{Team odds with home-ice advantage at home}

\item{away_odds}{Team odds with home-ice advantage at away (on the road)}

\item{home_win}{Number of home ice advantage team wins thus far in the series. Default to 0 (prediction before series start)}

\item{away_win}{Number of away team wins thus far in the series}

\item{ngames}{Number of games in the series, defaults to 7}

\item{game_home}{vector of T/F for 'home team' home games. Defaults for NHL best of 7 series: \code{c(T,T,F,F,T,F,T)}}

\item{predict_games_to_win}{(Defualt False) If TRUE, returns the table of ways the series could finish.}
}
\value{
numeric odds of home team win series (1-odds for away odds)
}
\description{
Given home and away win odds, produce the odds of the 'home advantage' team winning the series. From \url{http://www.stat.umn.edu/geyer/playoff.html}, modified to function with odds determination.
}
\references{
\url{http://www.stat.umn.edu/geyer/playoff.html}
}
