% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics-pace.R
\name{plot_pace_by_division}
\alias{plot_pace_by_division}
\title{Plot Pace By Division}
\usage{
plot_pace_by_division(
  graphic_dir = getOption("HockeyModel.graphics.path"),
  subdir = "pace",
  prediction_dir = getOption("HockeyModel.prediction.path"),
  scores = HockeyModel::scores
)
}
\arguments{
\item{graphic_dir}{Graphics Directory}

\item{subdir}{Subdirectory for pace graphics}

\item{prediction_dir}{Directory for predictions}

\item{scores}{HockeyModel::scores}
}
\value{
ggplot graphic of team pace vs. predicted
}
\description{
Plot Pace By Division
}
