% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics-pace.R
\name{plot_pace_by_team}
\alias{plot_pace_by_team}
\title{Plot Pace by team}
\usage{
plot_pace_by_team(
  graphic_dir = getOption("HockeyModel.graphics.path"),
  subdir = "pace",
  prediction_dir = getOption("HockeyModel.prediction.path"),
  scores = HockeyModel::scores
)
}
\arguments{
\item{graphic_dir}{The graphics directory}

\item{subdir}{The pace subdirectory in graphics}

\item{prediction_dir}{The predictions directory}

\item{scores}{The HockeyModel::scores object, or custom scores in the same format.}
}
\description{
Plot Pace by team
}
