% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics-plots.R
\name{plot_team_rating}
\alias{plot_team_rating}
\title{Plot Team Rating}
\usage{
plot_team_rating(m = HockeyModel::m, teamlist = NULL, league = "NHL")
}
\arguments{
\item{m}{HockeyModel::m. Pass \link{pwhl_m} for PWHL.}

\item{teamlist}{select a subset of teams, if desired}

\item{league}{League identifier. \code{"NHL"} (default) or \code{"PWHL"}. When
\code{"PWHL"}, PWHL team colours and title are used.}
}
\value{
a ggplot2 plot
}
\description{
Produces a plot of offensive and defensive ratings of teams, 0 centred (not scaled).
}
