% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley-terry.R
\name{plotBTGDModelFit}
\alias{plotBTGDModelFit}
\title{Plot the BT GD model Fit}
\usage{
plotBTGDModelFit(fittedBT = HockeyModel::fittedBT,
  dataBT = HockeyModel::dataBT)
}
\arguments{
\item{fittedBT}{the HockeyModel::fittedBT Model}

\item{dataBT}{the HockeyModel::dataBT data}
}
\value{
a baseplot from BradleyTerry2
}
\description{
Plot the BT GD model Fit
}
