% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley-terry (original).R, R/bradley-terry.R
\name{plotBTModelFit}
\alias{plotBTModelFit}
\title{Plot the BT model Fit}
\usage{
plotBTModelFit(fittedBT = HockeyModel::fittedBT,
  dataBT = HockeyModel::dataBT)

plotBTModelFit(fittedBT = HockeyModel::fittedBT,
  dataBT = HockeyModel::dataBT)
}
\arguments{
\item{fittedBT}{the HockeyModel::fittedBT Model}

\item{btdata}{the HockeyModel::dataBT data}

\item{fittedBT}{the HockeyModel::fittedBT Model}

\item{btdata}{the HockeyModel::dataBT data}
}
\value{
a baseplot from BradleyTerry2

a baseplot from BradleyTerry2
}
\description{
Plot the BT model Fit

Plot the BT model Fit
}
