% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{plotDC}
\alias{plotDC}
\title{Produce a plot of each team's offence and defence scores}
\usage{
plotDC(m = HockeyModel::m, teamlist = NULL)
}
\arguments{
\item{m}{m from data(m)}

\item{teamlist}{Teams to plot.}
}
\value{
a ggplot object
}
\description{
Produce a plot of each team's offence and defence scores
}
