% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.R
\name{predictEloResult}
\alias{predictEloResult}
\title{Calculate the win chance percent for HomeTeam}
\usage{
predictEloResult(home_rank, away_rank, h_adv = 0)
}
\arguments{
\item{home_rank}{The ranking of the Home Team.}

\item{away_rank}{The ranking of the Away Team.}

\item{h_adv}{The home advantage (in ranking points). Default: 0}
}
\value{
A number between 0 and 1 corresponding to the win chances of the Home Team.
}
\description{
Calculate the win chance percent for HomeTeam
}
\keyword{internal}
