% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-dixon-coles-workflow.R
\name{predictMultipleDaysResultsDC}
\alias{predictMultipleDaysResultsDC}
\title{Predict Multiple Days's W/L/D results}
\usage{
predictMultipleDaysResultsDC(
  startDate,
  endDate,
  schedule = HockeyModel::schedule,
  scores = HockeyModel::scores
)
}
\arguments{
\item{startDate}{First day of predicted results}

\item{endDate}{Last day of predicted results}

\item{schedule}{HockeyModel::schedule}

\item{scores}{HockeyModel::scores}
}
\value{
a data frame like HockeyModel::schedule with HomeWin, AwayWin and Draw odds
}
\description{
Calculate what each days' game predictions were by recalculating rho and m using games only up to the previous day. This is kinda slow, so a full season might take an hour or more depending on your hardware. Split out of the SeasonMetrics code to help with requests by @JB4991 on twitter.
}
