% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced.R
\name{prepareAdvancedData}
\alias{prepareAdvancedData}
\title{Prepare Advanced Data}
\usage{
prepareAdvancedData(advanced_scores = HockeyModel::advanced_scores,
  long = TRUE)
}
\arguments{
\item{advanced_scores}{an Advanced Scores frame from HockeyScrapR}

\item{long}{whether to return long (one team per row) or wide (both teams per row) data}
}
\value{
a data frame
}
\description{
Prepare Advanced Data
}
