% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwhl_beta}
\alias{pwhl_beta}
\title{PWHL \code{beta}}
\format{
a single numeric value, or \code{NULL} when not yet fitted
}
\usage{
pwhl_beta
}
\description{
Weibull shape parameter for PWHL tie-game enhancement. Populated by
\code{\link[=.update_dc_pwhl]{.update_dc_pwhl()}}.
}
\keyword{datasets}
