% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhl-model.R
\name{pwhl_daily_odds_table}
\alias{pwhl_daily_odds_table}
\title{PWHL daily odds table}
\usage{
pwhl_daily_odds_table(
  today = Sys.Date(),
  params = NULL,
  schedule = HockeyModel::pwhlSchedule
)
}
\arguments{
\item{today}{(\code{Date}) Date for games. Defaults to today.}

\item{params}{(\code{list} or \code{NULL}) PWHL DC parameter list.}

\item{schedule}{(\code{data.frame}) PWHL schedule.}
}
\value{
A \code{gt} table object, or \code{NULL}.
}
\description{
Returns a \code{gt} table of win odds and expected goals for today's
(or a specified date's) PWHL games. Returns \code{NULL} when there are no games.
}
