% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwhl_eta}
\alias{pwhl_eta}
\title{PWHL \code{eta}}
\format{
a single numeric value, or \code{NULL} when not yet fitted
}
\usage{
pwhl_eta
}
\description{
Weibull scale parameter for PWHL tie-game enhancement. Populated by
\code{\link[=.update_dc_pwhl]{.update_dc_pwhl()}}.
}
\keyword{datasets}
