% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhl-api-fetch.R
\name{pwhl_game_summary}
\alias{pwhl_game_summary}
\title{Fetch a single PWHL game summary}
\usage{
pwhl_game_summary(gid)
}
\arguments{
\item{gid}{(\code{integer(1)}) PWHL game ID.}
}
\value{
A one-row data frame or \code{NULL} if the game is not final.
}
\description{
Fetch a single PWHL game summary
}
\keyword{internal}
