% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhl-api-interface.R
\name{pwhl_get_long_team}
\alias{pwhl_get_long_team}
\title{Convert PWHL short team codes to full team names}
\usage{
pwhl_get_long_team(codes, pwhlTeamColours = HockeyModel::pwhlTeamColours)
}
\arguments{
\item{codes}{(\code{character}) PWHL team short code(s).}

\item{pwhlTeamColours}{(\code{data.frame}) PWHL team metadata table.}
}
\value{
(\code{character}) Full team name(s).
}
\description{
Convert PWHL short team codes to full team names
}
\keyword{internal}
