% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhl-api-interface.R
\name{pwhl_get_short_team}
\alias{pwhl_get_short_team}
\title{Convert PWHL full team names to short codes}
\usage{
pwhl_get_short_team(teams, pwhlTeamColours = HockeyModel::pwhlTeamColours)
}
\arguments{
\item{teams}{(\code{character}) Full PWHL team name(s).}

\item{pwhlTeamColours}{(\code{data.frame}) PWHL team metadata table.}
}
\value{
(\code{character}) Short team code(s).
}
\description{
Convert PWHL full team names to short codes
}
\keyword{internal}
