% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhl-model-core.R
\name{pwhl_in_season}
\alias{pwhl_in_season}
\title{Is there an active PWHL season?}
\usage{
pwhl_in_season(date = Sys.Date(), schedule = HockeyModel::pwhlSchedule)
}
\arguments{
\item{date}{(\code{Date}) Date to test. Defaults to today.}

\item{schedule}{(\code{data.frame}) PWHL schedule. Defaults to
\link{pwhlSchedule}.}
}
\value{
\code{logical(1)}.
}
\description{
Returns \code{TRUE} if \code{date} falls within the first and last game
dates recorded in \code{schedule}, i.e. the PWHL season has started but not
yet finished.
}
