% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwhl_m}
\alias{pwhl_m}
\title{PWHL Dixon-Coles model \code{m}}
\format{
a glm model fit, or \code{NULL} when not yet fitted
}
\usage{
pwhl_m
}
\description{
Fitted GLM capturing each PWHL team's attack and defence strength plus home
ice advantage. Populated by \code{\link[=.update_dc_pwhl]{.update_dc_pwhl()}}.
}
\keyword{datasets}
