% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhl-model.R
\name{pwhl_plot_odds_today}
\alias{pwhl_plot_odds_today}
\title{Plot today's PWHL game odds}
\usage{
pwhl_plot_odds_today(
  today = Sys.Date(),
  params = NULL,
  schedule = HockeyModel::pwhlSchedule
)
}
\arguments{
\item{today}{(\code{Date}) Date to predict. Defaults to today.}

\item{params}{(\code{list} or \code{NULL}) PWHL DC parameter list.}

\item{schedule}{(\code{data.frame}) PWHL schedule.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object, or \code{NULL}.
}
\description{
Produces a stacked bar chart of home/away/OT win odds for each
PWHL game scheduled on \code{today}. Returns \code{NULL} if there are no games.
}
