% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhl-model.R
\name{pwhl_plot_team_rating}
\alias{pwhl_plot_team_rating}
\title{PWHL team ratings plot}
\usage{
pwhl_plot_team_rating(m = HockeyModel::pwhl_m, teamlist = NULL)
}
\arguments{
\item{m}{(\code{glm}) PWHL DC model. Defaults to \link{pwhl_m}.}

\item{teamlist}{(\code{character}) Subset of teams. \code{NULL} uses all teams in \code{m}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Produces an offence-vs-defence scatter plot for PWHL teams
using the fitted PWHL model \code{m}.
}
