% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhl-model.R
\name{pwhl_remainder_season_dc}
\alias{pwhl_remainder_season_dc}
\title{Simulate the remainder of the PWHL regular season}
\usage{
pwhl_remainder_season_dc(scores, schedule, params = NULL)
}
\arguments{
\item{scores}{(\code{data.frame}) PWHL scores with a \code{Result} column.}

\item{schedule}{(\code{data.frame}) PWHL full-season schedule.}

\item{params}{(\code{list} or \code{NULL}) PWHL DC parameter list.}
}
\value{
A data frame of \code{HomeTeam}, \code{AwayTeam}, \code{HomeWin}, \code{AwayWin},
\code{Draw}, \code{GameID}, and \code{Date} for un-played games.
}
\description{
Produces a data frame of win odds for all remaining PWHL
regular-season games, used as input for \code{\link[=pwhl_loopless_sim]{pwhl_loopless_sim()}}.
}
\keyword{internal}
