% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhl-api-fetch.R
\name{pwhl_resolve_team_name}
\alias{pwhl_resolve_team_name}
\title{Resolve a PWHL team name from PWHLID}
\usage{
pwhl_resolve_team_name(pwhlID, pwhlTeamColours = HockeyModel::pwhlTeamColours)
}
\arguments{
\item{pwhlID}{the numeric pwhlID}

\item{pwhlTeamColours}{Built-in PWHL Team Colours data}
}
\value{
(\code{character(1)}) Canonical team name.
}
\description{
Determines the PWHL Team Name given the PWHLID from API calls
}
\keyword{internal}
