% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwhlSchedule}
\alias{pwhlSchedule}
\title{PWHL Schedule}
\format{
A data frame with variables:
\describe{
\item{Date}{Date the game is/was played}
\item{HomeTeam}{Name of the home team}
\item{AwayTeam}{Name of the visiting team}
\item{GameID}{Unique PWHL game ID}
\item{GameType}{\code{"R"} for regular season, \code{"P"} for playoff}
\item{GameStatus}{\code{"Scheduled"} or \code{"Final"}}
}
}
\source{
\url{https://www.thepwhl.com/}
}
\usage{
pwhlSchedule
}
\description{
Schedule for the current PWHL season (no scores included).
}
\keyword{datasets}
