% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwhlScores}
\alias{pwhlScores}
\title{PWHL Scores}
\format{
A data frame with variables:
\describe{
\item{Date}{Date the game was played}
\item{HomeTeam}{Name of the home team}
\item{AwayTeam}{Name of the visiting team}
\item{GameID}{Unique PWHL game ID}
\item{HomeGoals}{Goals scored by the home team}
\item{AwayGoals}{Goals scored by the away team}
\item{OTStatus}{Overtime status: \code{""} (regulation), \code{"OT"}, or \code{"SO"}}
\item{GameType}{\code{"R"} for regular season, \code{"P"} for playoff}
\item{GameStatus}{\code{"Final"}}
}
}
\source{
\url{https://www.thepwhl.com/}
}
\usage{
pwhlScores
}
\description{
Historical PWHL game scores.
}
\keyword{datasets}
