% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwhlTeamColours}
\alias{pwhlTeamColours}
\title{PWHL Team Colours}
\format{
A data frame with variables:
\describe{
\item{Team}{Full team name}
\item{Hex}{Primary colour hex code}
\item{AltHex}{Secondary colour hex code}
\item{Hashtag}{Official social media hashtag}
\item{ShortCode}{Three-letter (or two-letter) team abbreviation}
\item{Division}{League division (currently all \code{"PWHL"})}
\item{Conference}{League conference (currently all \code{"PWHL"})}
\item{PWHLID}{Integer team ID used by the HockeyTech API}
}
}
\source{
\url{https://www.thepwhl.com/}
}
\usage{
pwhlTeamColours
}
\description{
Primary and secondary hex colours, team short codes, hashtags, and PWHL team
IDs for all six PWHL teams.
}
\keyword{datasets}
