% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_iterative.R
\name{readIterativePredictions}
\alias{readIterativePredictions}
\title{Load iterative rankings and parameters from disk}
\usage{
readIterativePredictions(
  filename,
  filepath = "./prediction_results/iterative_ranking"
)
}
\arguments{
\item{filename}{(\code{character(1)}) File name to read.}

\item{filepath}{(\code{character(1)}) Directory containing saved files.}
}
\value{
(\code{list}) Saved \code{Rankings} and \code{Params}.
}
\description{
Load iterative rankings and parameters from disk
}
\keyword{internal}
