% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{recovery}
\alias{recovery}
\title{Recovery Calculating}
\usage{
recovery(pPoints, sdPoints, aPoints)
}
\arguments{
\item{pPoints}{vector of predcited points}

\item{sdPoints}{vector of predicted points standard deviations}

\item{aPoints}{vector of actual points}
}
\value{
single value 0-31
}
\description{
Recovery Calculating
}
