% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley-terry.R
\name{remainderSeasonBT}
\alias{remainderSeasonBT}
\title{BT remainder of season}
\usage{
remainderSeasonBT(nsims = 10000, scores = HockeyModel::scores,
  schedule = HockeyModel::schedule, fittedBT = HockeyModel::fittedBT,
  odds = FALSE, ...)
}
\arguments{
\item{nsims}{Number of simulations}

\item{scores}{the historical scores}

\item{schedule}{uplayed future games}

\item{odds}{Whether to return raw odds (TRUE) or go to simulation.}

\item{...}{additional parameters to pass to season simulation (in parallel)}
}
\value{
data frame of Team, playoff odds.
}
\description{
Odds for each team to get to playoffs.
}
