% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.R
\name{remainderSeasonELO}
\alias{remainderSeasonELO}
\title{ELO remainder of season}
\usage{
remainderSeasonELO(nsims = 10000, scores = HockeyModel::scores,
  schedule = HockeyModel::schedule, odds = FALSE,
  elos = HockeyModel::elos, ...)
}
\arguments{
\item{nsims}{Number of simulations}

\item{scores}{the historical scores}

\item{schedule}{uplayed future games}

\item{odds}{whether to return odds table or predictions}

\item{elos}{HockeyModel::elos or custom value}

\item{...}{arguements to pass to elo predictor & simulateSeason}
}
\value{
data frame of Team, playoff odds.
}
\description{
Odds for each team to get to playoffs.
}
