% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rho}
\alias{rho}
\title{Dixon Coles' \code{rho}}
\format{
a single numerical value usually around -0.25
}
\usage{
rho
}
\description{
Rho for low score correction to Dixon Coles. Should be around -0.25
}
\keyword{datasets}
