% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rmse}
\alias{rmse}
\title{RMSE}
\usage{
rmse(predicted, actual)
}
\arguments{
\item{predicted}{Predicted numeric value}

\item{actual}{Actual numeric value}
}
\value{
a single value for RMSE
}
\description{
calculate the RMSE metrics. From MLMetrics
}
