% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rsquare}
\alias{rsquare}
\title{R Square}
\usage{
rsquare(predicted, actual)
}
\arguments{
\item{predicted}{Predicted numeric value}

\item{actual}{Actual numeric value}
}
\value{
a single value for R^2
}
\description{
calculate the R^2 metrics. From MLMetrics
}
