% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_iterative.R
\name{saveIterativePredictions}
\alias{saveIterativePredictions}
\title{Save iterative rankings and parameters to disk}
\usage{
saveIterativePredictions(
  rankings,
  params,
  filepath = "./prediction_results/iterative_ranking"
)
}
\arguments{
\item{rankings}{(\code{data.frame}) Team rating table to save.}

\item{params}{(\code{list}) Iterative model parameters to save.}

\item{filepath}{(\code{character(1)}) Directory for output files.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Save iterative rankings and parameters to disk
}
\keyword{internal}
