% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schedule}
\alias{schedule}
\title{Schedule for current season.}
\format{A data frame with  variables:
\describe{
  \item{Date}{Date of the game played/to be played}
  \item{AwayTeam}{Name of the visiting team}
  \item{HomeTeam}{Name of the home team}
}}
\source{
\url{http://www.hockey-reference.com/}
}
\usage{
schedule
}
\description{
A dataframe with the schedule (no scores included) for current season.
}
\keyword{datasets}
