% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley-terry.R
\name{simpleBTPredict}
\alias{simpleBTPredict}
\title{Simple BT Predict Game}
\usage{
simpleBTPredict(btModel, home.team, away.team)
}
\arguments{
\item{away.team}{}

\item{btSimple}{simple Bradley Terry Model}
}
\description{
Simple BT Predict Game
}
