% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{simulateSeason}
\alias{simulateSeason}
\title{Simulate the remainder of the season}
\usage{
simulateSeason(
  odds_table,
  scores = HockeyModel::scores,
  nsims = 10000,
  schedule = HockeyModel::schedule,
  progress = FALSE
)
}
\arguments{
\item{odds_table}{A dataframe with HomeTeam, AwayTeam, HomeWin, AwayWin, Draw, and Date}

\item{scores}{Past (historical) season scores. Defaults to HockeyModel::Scores}

\item{nsims}{number of simulations to run}

\item{schedule}{Future unplayed games. Defaults to HockeyModel::schedule}

\item{progress}{whether to show a progress bar.}
}
\value{
a data frame of results
}
\description{
Simulate the remainder of the season
}
