% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{solvePlayoffsMC}
\alias{solvePlayoffsMC}
\title{simulate Playoffs}
\usage{
solvePlayoffsMC(
  summary_results,
  nsims = 1e+06,
  cores = parallel::detectCores() - 1
)
}
\arguments{
\item{summary_results}{summary results}

\item{nsims}{Number of playoff sims to run. Too many takes a long time.}

\item{cores}{Number of processor cores to use}
}
\value{
a data frame of each teams' odds of winning each round (First Round, Second Round, Conference Finals and Stanley Cup)
}
\description{
Solves playoff odds by MC simulation.
}
