% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.R
\name{splitDates}
\alias{splitDates}
\title{Split dates to by season if multiple seasons are calculated together}
\usage{
splitDates(game_dates, season_split = "-08-01")
}
\arguments{
\item{game_dates}{The dates of games to be split by season as a vector of Dates, or as a df with dates in game_dates$Date.}

\item{season_split}{The annual date by which to split seasons. As string '-MM-DD' format. For NHL, August 1 is chosen, so '-08-01'}
}
\value{
A list of vectors of dates.
}
\description{
Split dates to by season if multiple seasons are calculated together
}
\keyword{internal}
