% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summary_results_testing}
\alias{summary_results_testing}
\title{Summary Results (Testing)
A testing dataset from 2021 presesason. Not for regular use.}
\format{
a data frame
}
\usage{
summary_results_testing
}
\description{
Summary Results (Testing)
A testing dataset from 2021 presesason. Not for regular use.
}
\keyword{datasets}
