% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-dixon-coles-fit.R
\name{tau}
\alias{tau}
\title{Tau function}
\usage{
tau(xx, yy, lambda, mu, rho)
}
\arguments{
\item{xx}{Homeoal value for adjustment factor calculation}

\item{yy}{Away Goal value for adjustment factor calculation}

\item{lambda}{Home goal expected for adjustment factor calculation}

\item{mu}{Away Goal expected for adjustment factor calculation}

\item{rho}{the factor for adjustment calculations}
}
\description{
Used in Dixon-Coles's to adjust low scores
}
