% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{team_progression_odds}
\alias{team_progression_odds}
\title{Team Progression Odds}
\usage{
team_progression_odds(round, team, odds)
}
\arguments{
\item{round}{Playoff round (1-4)}

\item{team}{Team to calculate odds for}

\item{odds}{input odds of each opponent team making each position of the round}
}
\value{
a vector of odds of progressing in each position
}
\description{
Team Progression Odds
}
