% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teamColours}
\alias{teamColours}
\title{Team Colours
Hex and RGB team colours (primary & secondary where applicable), and path to team logos}
\format{a data frame of colours & logos}
\usage{
teamColours
}
\description{
Team Colours
Hex and RGB team colours (primary & secondary where applicable), and path to team logos
}
\keyword{datasets}
