% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theta}
\alias{theta}
\title{Theta}
\format{
a single numerical value
}
\usage{
theta
}
\description{
Theta for adjustment of tie games in the model. Theta is the 'lambda' value of a Poisson distribution, optimized to fit the ties data. Should be around 2.5
}
\keyword{datasets}
