% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.R
\name{todayELO}
\alias{todayELO}
\title{DC Predictions Today}
\usage{
todayELO(today = Sys.Date(), elos = HockeyModel::elos,
  schedule = HockeyModel::schedule)
}
\arguments{
\item{today}{Generate predictions for this date. Defaults to today}

\item{elos}{Elo History (to extract last value)}

\item{schedule}{shcedule to use, if not the builtin}
}
\value{
a data frame of HomeTeam, AwayTeam, HomeWin, AwayWin, Draw, or NULL if no games today
}
\description{
DC Predictions Today
}
