% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-league-simulation.R
\name{todayOdds}
\alias{todayOdds}
\title{Today's Odds}
\usage{
todayOdds(
  params = NULL,
  today = Sys.Date(),
  schedule = HockeyModel::schedule,
  expected_mean = NULL,
  season_percent = NULL,
  include_xG = FALSE
)
}
\arguments{
\item{params}{The named list containing m, rho, beta, eta, and k. See \code{\link[=updateModel]{updateModel()}} for information on the params list}

\item{today}{The date for which you want game odds}

\item{schedule}{The schedule, default to internal schedule}

\item{expected_mean}{the mean lambda & mu, used only for regression}

\item{season_percent}{the percent complete of the season, used for regression}

\item{include_xG}{Whether to include xG values in reported odds}
}
\value{
a data frame of HomeTeam, AwayTeam, HomeWin, AwayWin, Draw, or NULL if no games today
}
\description{
Determine today's games' odds (if today has games), or a specified date's odds
}
\examples{
todayOdds(today = as.Date("2019-11-01"))
}
