% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeForecasted_PythagoreanPuck.R
\name{todayPP}
\alias{todayPP}
\title{PythagoreanPuck Predictions Today}
\usage{
todayPP(today = Sys.Date(), schedule = HockeyModel::schedule)
}
\arguments{
\item{today}{Generate predictions for this date. Defaults to today}

\item{schedule}{shcedule to use, if not the builtin}
}
\value{
a data frame of HomeTeam, AwayTeam, HomeWin, AwayWin, Draw, or NULL if no games today
}
\description{
PythagoreanPuck Predictions Today
}
