% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-social.R
\name{tweetMetrics}
\alias{tweetMetrics}
\title{Tweet Metrics}
\usage{
tweetMetrics()
}
\value{
(\code{data.frame}) A summary of the attempted post (see
\code{.summarize_post_results()}), invisibly.
}
\description{
Tweet the metrics (Log Loss and Accuracy)
}
