% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-social.R
\name{tweetPace}
\alias{tweetPace}
\title{Tweet Pace Plots}
\usage{
tweetPace(
  delay = stats::runif(1, min = 1, max = 3) * 60,
  graphic_dir = getOption("HockeyModel.graphics.path"),
  subdir = "pace",
  prediction_dir = getOption("HockeyModel.prediction.path"),
  scores = HockeyModel::scores
)
}
\arguments{
\item{delay}{Delay between posted tweets}

\item{graphic_dir}{The graphics directory}

\item{subdir}{The pace subdirectory in graphics}

\item{prediction_dir}{The predictions directory}

\item{scores}{HockeyModel::scores or a custom value}
}
\value{
(\code{data.frame}) A summary of each attempted post (see
\code{.summarize_post_results()}), invisibly.
}
\description{
Tweet Pace Plots
}
