% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_iterative.R
\name{updateRankingsToToday}
\alias{updateRankingsToToday}
\title{Replay completed games to update iterative rankings state}
\usage{
updateRankingsToToday(
  rankings,
  params,
  scores = HockeyModel::scores,
  rankings_date = getSeasonStartDate()
)
}
\arguments{
\item{rankings}{(\code{data.frame}) Starting rankings.}

\item{params}{(\code{list}) Iterative parameters used for updates.}

\item{scores}{(\code{data.frame}) Historical game results.}

\item{rankings_date}{(\code{Date}) Earliest date to include for replay.}
}
\value{
(\code{data.frame}) Updated rankings.
}
\description{
Replay completed games to update iterative rankings state
}
\keyword{internal}
