% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{updateScores}
\alias{updateScores}
\title{Update Scores}
\usage{
updateScores(data_dir = "./data-raw", last_playoffs = FALSE)
}
\arguments{
\item{data_dir}{The data storage directory}

\item{last_playoffs}{Trigger to pass to \code{\link[HockeyScrapR]{updateScores}}}
}
\value{
score
}
\description{
DEPRECIATED: USE \code{\link{updateScoresAPI}} Updates the scores of all finished games.
}
