% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interface.R
\name{updateScoresAPI_byGameID}
\alias{updateScoresAPI_byGameID}
\title{Update scores for specific game IDs}
\usage{
updateScoresAPI_byGameID(gameids, save_data = FALSE)
}
\arguments{
\item{gameids}{(\code{character} or \code{numeric}) One or more NHL game IDs.}

\item{save_data}{(\code{logical(1)}) Whether to write updated scores into package
data when \code{usethis} is available.}
}
\value{
(\code{scores}) Invisibly, the updated scores data frame.
}
\description{
Update scores for specific game IDs
}
\keyword{internal}
