% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{updateSeries}
\alias{updateSeries}
\title{Update Series}
\usage{
updateSeries(series = NULL)
}
\arguments{
\item{series}{series, if provided, else update function.}
}
\value{
series df.
}
\description{
Updates the saved series data as manually adjusted in the df in the function
}
