% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R, R/utils.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy Calculating}
\usage{
accuracy(predicted, actual)

accuracy(predicted, actual)
}
\arguments{
\item{predicted}{Predicted odds of an event occuring. needen't be of set {0,1}}

\item{actual}{If the event occured (0 or 1), or model results in 0, 0.25, 0.4, 0.6, 0.75, 1.0}

\item{pWin}{vector of predcited win}

\item{result}{vector of actual win}
}
\value{
single value 0-1

a percentage of correct predictions
}
\description{
Accuracy Calculating

Accuracy Calculator
}
