% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy Calculating}
\usage{
accuracy(pWin, result)
}
\arguments{
\item{pWin}{vector of predcited win}

\item{result}{vector of actual win}
}
\value{
single value 0-1
}
\description{
Accuracy Calculating
}
