% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_postponed_to_schedule_end}
\alias{add_postponed_to_schedule_end}
\title{Add Postponed Games to Schedule End}
\usage{
add_postponed_to_schedule_end(schedule = HockeyModel::schedule)
}
\arguments{
\item{schedule}{the schedule to reconfigure}
}
\value{
a schedule with postponed games moved to the end of the schedule - helps to not drop games that are otherwise in the past but weren't played.
}
\description{
Sometimes games are postponed without a makeup date initially announced. The model just drops those games if this function is not employed to move the games to the end of the season
Note that the games are all dumped on one day so it doesn't account for back to back or travel days or anything.
}
