% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addSeriesWin}
\alias{addSeriesWin}
\title{Add Series Win}
\usage{
addSeriesWin(series, winner)
}
\arguments{
\item{series}{the full series dataframe}

\item{winner}{Team to add a win to}
}
\value{
updated series
}
\description{
Add Series Win
}
