% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{advanced_scores}
\alias{advanced_scores}
\title{Historical NHL scores with Advanced Stats}
\format{
A data frame with  variables
}
\source{
\url{http://www.hockey-reference.com/}
}
\usage{
advanced_scores
}
\description{
A dataset containing historical NHL and WHA scores.
}
\keyword{datasets}
