% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{cleanupPredictionsFile}
\alias{cleanupPredictionsFile}
\title{Cleanup Predictions File}
\usage{
cleanupPredictionsFile(
  filepath = file.path(devtools::package_file(), "data-raw", "dailyodds.csv")
)
}
\arguments{
\item{filepath}{file path to cleanup}
}
\description{
Sometimes the predictions file may end up with game duplicates (last minute postponements, etc,) This deduplicates, taking only the latest instance of a prediction (Games are unique by GameID)
}
