% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{daily_odds_table}
\alias{daily_odds_table}
\title{Today Odds Table}
\usage{
daily_odds_table(
  today = Sys.Date(),
  rho = HockeyModel::rho,
  m = HockeyModel::m,
  schedule = HockeyModel::schedule
)
}
\arguments{
\item{today}{A date for games to create a table. Defaults to today.}

\item{rho}{Hockey Model rho}

\item{m}{Hockey Model m}

\item{schedule}{Schedule, or HockeyModel Schedule}
}
\value{
a gt table
}
\description{
Returns a gt table of odds for today's games (or games for a supplied date)
}
