% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dailySummary}
\alias{dailySummary}
\title{Daily functions, rolled into one call}
\usage{
dailySummary(graphic_dir = "./prediction_results/graphics/", ...)
}
\arguments{
\item{...}{Items to pass to update and plot functions.}
}
\description{
Daily functions, rolled into one call
}
