% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dailySummary}
\alias{dailySummary}
\title{Daily functions, rolled into one call}
\usage{
dailySummary(
  graphic_dir = "./prediction_results/graphics",
  subdir = "pace",
  twitter_token = NULL,
  delay = stats::runif(1, min = 2, max = 6) * 60
)
}
\arguments{
\item{graphic_dir}{Directory for graphic files}

\item{subdir}{subdirectory to \code{graphic_dir} for pace plots}

\item{delay}{delay between tweet posts}

\item{token}{token to pass to rtweet calls}
}
\description{
Daily functions, rolled into one call
}
