% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dailySummary}
\alias{dailySummary}
\title{Daily functions, rolled into one call}
\usage{
dailySummary(
  graphic_dir = "./prediction_results/graphics/",
  token = rtweet::get_token(),
  delay = stats::runif(1, min = 4, max = 8) * 60
)
}
\arguments{
\item{graphic_dir}{Directory for graphic files}

\item{token}{token to pass to rtweet calls}

\item{delay}{delay between tweet posts}
}
\description{
Daily functions, rolled into one call
}
