% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dailySummary}
\alias{dailySummary}
\title{Daily functions, rolled into one call}
\usage{
dailySummary(
  graphic_dir = "./prediction_results/graphics/",
  token = rtweet::get_token(),
  delay = 60 * 10,
  ...
)
}
\arguments{
\item{graphic_dir}{Directory for graphic files}

\item{token}{token to pass to rtweet calls}

\item{delay}{delay between tweet posts}

\item{...}{Items to pass to update and plot functions.}
}
\description{
Daily functions, rolled into one call
}
