% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DC_XI_NHL}
\alias{DC_XI_NHL}
\title{Default NHL Dixon-Coles sigmoid time-decay slope}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DC_XI_NHL
}
\description{
\code{xi} controls how steeply the within-season logistic weighting curve drops.
}
\keyword{internal}
