% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{DCPredict}
\alias{DCPredict}
\title{DC Predict home/draw/away win}
\usage{
DCPredict(home, away, m = NULL, rho = NULL, maxgoal = 8,
  scores = scores)
}
\arguments{
\item{home}{home team}

\item{away}{away team}

\item{m}{result from getM}

\item{rho}{rho from getRho}

\item{maxgoal}{max number of goals per team}

\item{scores}{optional, if not supplying m & rho, scores used to calculate them.}
}
\value{
a list of home win, draw, and away win probability
}
\description{
Using Dixon Coles technique, predict odds each of home win, draw, and away win.
}
\keyword{internal}
