% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{dcPredictMultipleDays}
\alias{dcPredictMultipleDays}
\title{DC Predict Multiple Days}
\usage{
dcPredictMultipleDays(
  start = as.Date(getSeasonStartDate()),
  end = Sys.Date(),
  scores = HockeyModel::scores,
  schedule = HockeyModel::schedule,
  filedir = "./prediction_results"
)
}
\arguments{
\item{start}{First day to predict. Default start of season}

\item{end}{Last day to predict. Default today}

\item{scores}{HockeyModel::scores}

\item{schedule}{HockeyModel::schedule}

\item{filedir}{Where to save prediction files.}
}
\value{
true, if successful
}
\description{
For catching up on daily predictions
}
