% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{dcProbMatrix}
\alias{dcProbMatrix}
\title{Generate the Dixon-Coles' Probability Matrix}
\usage{
dcProbMatrix(
  home,
  away,
  m = HockeyModel::m,
  rho = HockeyModel::rho,
  maxgoal = 8,
  scores = HockeyModel::scores,
  expected_mean = NULL,
  season_percent = NULL
)
}
\arguments{
\item{home}{home team}

\item{away}{away team}

\item{m}{result from getM}

\item{rho}{rho from getRho}

\item{maxgoal}{max number of goals per team}

\item{scores}{optional, if not supplying m & rho, scores used to calculate them.}

\item{expected_mean}{the mean lambda & mu, used only for regression}

\item{season_percent}{the percent complete of the season, used for regression}
}
\value{
a square matrix of dims 0:maxgoal with odds at each count of  home goals on 'rows' and away goals  on 'columns'
}
\description{
Generate the Dixon-Coles' Probability Matrix
}
