% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{dcRealSeasonPredict}
\alias{dcRealSeasonPredict}
\title{DC Simulate season with reevaluation every n days}
\usage{
dcRealSeasonPredict(nsims = 1e+05, scores = HockeyModel::scores,
  schedule = HockeyModel::schedule, cores = parallel::detectCores() -
  1, ndays = 7, regress = TRUE)
}
\arguments{
\item{nsims}{number of simulations}

\item{scores}{hisotircal scores}

\item{schedule}{schedule this season}

\item{cores}{number of cores to process}

\item{ndays}{number of days of games to play before reevaluating m}

\item{regress}{Whether to regress toward mean}
}
\value{
data frame of Team point ranges, playoff odds, etc.
}
\description{
DC Simulate season with reevaluation every n days
}
