% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-dixon-coles-predict.R
\name{dcResult}
\alias{dcResult}
\title{DC Result Sample}
\usage{
dcResult(lambda, mu, params = NULL, maxgoal = 8, nsim = 1)
}
\arguments{
\item{lambda}{home team lambda}

\item{mu}{away team mu}

\item{params}{The named list containing m, rho, beta, eta, and k. See \link{updateDC} for information on the params list}

\item{maxgoal}{max goals predicable per game, default 10}

\item{nsim}{the number of simulations in each result}
}
\value{
a result from 0 to 1 corresponding to \link{scores} results
}
\description{
DC Result Sample
}
