% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{dcResult}
\alias{dcResult}
\title{DC Result Sample}
\usage{
dcResult(lambda, mu, rho = HockeyModel::rho, maxgoal = 10)
}
\arguments{
\item{lambda}{home team lambda}

\item{mu}{away team mu}

\item{rho}{HockeyModel::rho}

\item{maxgoal}{max goals predicable per game, default 10}
}
\value{
a result from 0 to 1 corresponding to \link{score} results
}
\description{
DC Result Sample
}
