% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{dcResult}
\alias{dcResult}
\title{DC Result Sample}
\usage{
dcResult(
  lambda,
  mu,
  rho = HockeyModel::rho,
  theta = HockeyModel::theta,
  gamma = HockeyModel::gamma,
  maxgoal = 10
)
}
\arguments{
\item{lambda}{home team lambda}

\item{mu}{away team mu}

\item{rho}{HockeyModel::rho}

\item{theta}{HockeyModel::theta}

\item{gamma}{HockeyModel::gamma}

\item{maxgoal}{max goals predicable per game, default 10}
}
\value{
a result from 0 to 1 corresponding to \link{scores} results
}
\description{
DC Result Sample
}
