% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-dixon-coles-predict.R
\name{dcSample}
\alias{dcSample}
\title{DC Sample}
\usage{
dcSample(
  home,
  away,
  params = NULL,
  maxgoal = 10,
  scores = HockeyModel::scores,
  expected_mean = NULL,
  season_percent = NULL,
  as_result = TRUE
)
}
\arguments{
\item{home}{home team}

\item{away}{away team}

\item{params}{The named list containing m, rho, beta, eta, and k. See \link{updateDC} for information on the params list}

\item{maxgoal}{max number of goals per team}

\item{scores}{optional, if not supplying m & rho, scores used to calculate them.}

\item{expected_mean}{the mean lambda & mu, used only for regression}

\item{season_percent}{the percent complete of the season, used for regression}

\item{as_result}{Whether to give a score or just flatten it to result type (see \link{scores} for score result)}
}
\value{
a random Home & away goals & OT/SO status if needed
}
\description{
Get a random single game result using DC method. repeated running should give a new value each time
}
\examples{
dcSample("Toronto Maple Leafs", "Montreal Canadiens")
}
