% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{DCweights}
\alias{DCweights}
\title{Compute sigmoid time-decay weights for historical matches}
\usage{
DCweights(dates, currentDate = Sys.Date(), xi = 0.01, upsilon = 365.25)
}
\arguments{
\item{dates}{(\code{Date}) Match dates.}

\item{currentDate}{(\code{Date}) Reference date for weighting.}

\item{xi}{(\code{double(1)}) Logistic slope.}

\item{upsilon}{(\code{double(1)}) Logistic midpoint.}
}
\value{
(\code{double}) Weight per input date.
}
\description{
Compute sigmoid time-decay weights for historical matches
}
\keyword{internal}
