% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{.daily_summary_nhl}
\alias{.daily_summary_nhl}
\title{Daily functions, rolled into one call}
\usage{
.daily_summary_nhl(
  graphic_dir = "./prediction_results/graphics",
  subdir = "pace",
  delay = stats::runif(1, min = 2, max = 6) * 60
)
}
\arguments{
\item{graphic_dir}{Directory for graphic files}

\item{subdir}{subdirectory to \code{graphic_dir} for pace plots}

\item{delay}{delay between tweet posts}
}
\value{
\code{NULL} (invisibly).
}
\description{
Daily functions, rolled into one call
}
\keyword{internal}
