% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.onAttach}
\alias{.onAttach}
\title{Print package startup message on attach}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{(\code{character(1)}) Library path provided by R.}

\item{pkgname}{(\code{character(1)}) Package name provided by R.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Print package startup message on attach
}
\keyword{internal}
