% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.onLoad}
\alias{.onLoad}
\title{Initialize package options on load}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{(\code{character(1)}) Library path provided by R.}

\item{pkgname}{(\code{character(1)}) Package name provided by R.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Initialize package options on load
}
\keyword{internal}
