% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{.playoff_odds_pwhl}
\alias{.playoff_odds_pwhl}
\title{Build the PWHL playoff-odds plot}
\usage{
.playoff_odds_pwhl(
  data_dir = getOption("HockeyModel.prediction.path", "./prediction_results")
)
}
\arguments{
\item{data_dir}{(\code{character(1)}) Root prediction directory.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Build the PWHL playoff-odds plot
}
\keyword{internal}
