% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{.today_odds_plot_nhl}
\alias{.today_odds_plot_nhl}
\title{Build the NHL today-odds plot}
\usage{
.today_odds_plot_nhl(
  date = Sys.Date(),
  params = NULL,
  schedule = HockeyModel::schedule,
  scores = HockeyModel::scores
)
}
\arguments{
\item{date}{(\code{Date}) Target date.}

\item{params}{(\code{list} or \code{NULL}) NHL model parameters.}

\item{schedule}{(\code{data.frame}) NHL schedule.}

\item{scores}{(\code{data.frame}) NHL scores.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object or \code{NULL}.
}
\description{
Build the NHL today-odds plot
}
\keyword{internal}
