% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{.update_model_nhl}
\alias{.update_model_nhl}
\title{Update the NHL front-end model payload}
\usage{
.update_model_nhl(save_data = TRUE)
}
\arguments{
\item{save_data}{(\code{logical(1)}) Whether to persist refreshed package data.}
}
\value{
(\code{list}) NHL \code{scores}, \code{schedule}, and \code{params}.
}
\description{
Update the NHL front-end model payload
}
\keyword{internal}
