% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{.update_predictions_nhl}
\alias{.update_predictions_nhl}
\title{Update saved NHL prediction snapshots}
\usage{
.update_predictions_nhl(
  data_dir = getOption("HockeyModel.prediction.path", "./prediction_results"),
  scores = HockeyModel::scores,
  schedule = HockeyModel::schedule,
  params = NULL
)
}
\arguments{
\item{data_dir}{(\code{character(1)}) Directory of saved NHL prediction files.}

\item{scores}{(\code{data.frame}) NHL scores.}

\item{schedule}{(\code{data.frame}) NHL schedule.}

\item{params}{(\code{list} or \code{NULL}) NHL model parameters.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Update saved NHL prediction snapshots
}
\keyword{internal}
