% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{format_playoff_odds}
\alias{format_playoff_odds}
\title{Format Playoff Odds}
\usage{
format_playoff_odds(
  playoff_odds,
  caption_text = "",
  trim = TRUE,
  trimcup = FALSE
)
}
\arguments{
\item{playoff_odds}{a playoff odds data frame with columns Team, Make_Playoffs, Win_First_Round, Win_Second_Round, Win_Conference, Win_Cup}

\item{caption_text}{Additional text to prepend to " Playoff Odds" in table title. E.g. 'Eastern Conference' if only eastteams sent in.}

\item{trim}{Whether to drop teams that have 0 chance of making playoffs. Default true}

\item{trimcup}{Whether to drorp teams that have 0 chance of winning cup. Default false}
}
\value{
a gt table
}
\description{
Takes a playoff odds table and returns a gt table
}
