% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interface.R
\name{get_xg}
\alias{get_xg}
\title{Get xG for one or many gameIds}
\usage{
get_xg(gameIds)
}
\arguments{
\item{gameIds}{one or many game Ids}
}
\value{
a list with Game ID, HomexG and AwayxG if one game ID supplied, or a data frame with those columns
}
\description{
Get xG for one or many gameIds
}
