% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interface.R
\name{getAPISeries}
\alias{getAPISeries}
\title{Get Playoff Series using the NHL API}
\usage{
getAPISeries(season = getCurrentSeason8())
}
\arguments{
\item{season}{Optional, the season's playoff series to retrieve}
}
\value{
a data frame with Round, Series, Home and Away Teams, number of wins each, playoff ranking/seed and whether the series is complete
}
\description{
Gets the current season (or previous seasons') playoff series information using the NHL API
}
