% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interface.R
\name{getCurrentSeason8}
\alias{getCurrentSeason8}
\title{Get Current Season}
\usage{
getCurrentSeason8()
}
\value{
current season (as 20172018 format) based on today's date
}
\description{
Get Current Season
}
