% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_iterative.R
\name{getIterativePredictions}
\alias{getIterativePredictions}
\title{Get Iterative Predictions}
\usage{
getIterativePredictions(
  date,
  schedule = HockeyModel::schedule,
  scores = HockeyModel::scores,
  rankings = HockeyModel::iterativeRankings,
  params = HockeyModel::iterativeParameters
)
}
\arguments{
\item{date}{Game date to get results for}

\item{schedule}{The HockeyModel::schedule object}

\item{scores}{The HockeyModel::scores object}

\item{rankings}{The HockeyModel::iterativeRankings object}

\item{params}{The HockeyModel:::iterativeParameters object}
}
\value{
a data.frame with colums for GameID, Home & Away Teams, Home & Away win percentage, and TotalxG predicted
}
\description{
Get predictions of all games on a date. Return win/loss and total xG from the Dixon-Coles iterative model
}
