% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{getM}
\alias{getM}
\title{Fast Dixon-Coles model fitting 'm'.}
\usage{
getM(scores = scores)
}
\arguments{
\item{scores}{the historical scores to evaluate}
}
\value{
a model 'm' of dixon coles type parameters.
}
\description{
Produces a
}
\keyword{internal}
