% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interface.R
\name{getNHLSchedule}
\alias{getNHLSchedule}
\title{Get NHL Schedule}
\usage{
getNHLSchedule(season = getSeason())
}
\arguments{
\item{season}{Season(s) for which the schedule is requested, as YYYY format. Defaults to current season. Accepts single years, ranges e.g.c(2010:2015) or complex ranges e.g. c(2010:2015, 2018).}
}
\value{
a data frame of all scheduled games for the season(s) requested, with Date, HomeTeam, AwayTeam, GameID, and GameType.
}
\description{
Gets the NHL schedule from the NHL API for the season(s) requested. Returns data formatted for further use. This can be slow if requesting many seasons due to the API rate limit.
}
\seealso{
See [nhlapi::nhl_schedule_seasons()] for alternative schedule requests, and [nhlapi::nhl_games()] for more information on GameID
}
