% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhl-api-interface.R
\name{getPWHLSchedule}
\alias{getPWHLSchedule}
\title{Get PWHL Schedule}
\usage{
getPWHLSchedule(
  season = getCurrentPWHLSeason(),
  pwhlTeamColours = HockeyModel::pwhlTeamColours
)
}
\arguments{
\item{season}{(\code{integer(1)}) PWHL season ID. Defaults to the current season
via \code{getCurrentPWHLSeason()}.}

\item{pwhlTeamColours}{(\code{data.frame}) The built-in PWHL team colours dataset
if not otherwise provided.}
}
\value{
A data frame with columns \code{Date}, \code{HomeTeam}, \code{AwayTeam}, \code{GameID},
\code{GameType}, and \code{GameStatus}.
}
\description{
Gets the PWHL game schedule for the season requested. Returns
a data frame formatted for use with the rest of the HockeyModel package.
}
