% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{getSeason}
\alias{getSeason}
\title{Get Season from Game Date}
\usage{
getSeason(gamedate)
}
\arguments{
\item{gamedate}{The date of the game to check for season}
}
\value{
a character season id (e.g. 20172018)
}
\description{
Get Season from Game Date
}
