% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interface.R
\name{getSeasonEndDate}
\alias{getSeasonEndDate}
\title{GetCurrentSeasonEndDate}
\usage{
getSeasonEndDate(season = NULL)
}
\arguments{
\item{season}{Season (8 character code) or NULL for current/most recent season}
}
\value{
Season end date (as date)
}
\description{
GetCurrentSeasonEndDate
}
