% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{getSeasonMetricsDC}
\alias{getSeasonMetricsDC}
\title{Get Season Metrics}
\usage{
getSeasonMetricsDC(schedule = HockeyModel::schedule,
  scores = HockeyModel::scores)
}
\arguments{
\item{schedule}{HockeyModel::schedule used to help with calculations}

\item{scores}{HockeyModel::scores used to compare to predictions}
}
\value{
a list of log loss and accuracy for the season
}
\description{
Calculates the Log Loss and Accuracy of the model by re-estimating m and rho daily and creating game odds
}
