% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{getSeasonMetricsDC}
\alias{getSeasonMetricsDC}
\title{Get Season Metrics}
\usage{
getSeasonMetricsDC()
}
\value{
a list of log loss and accuracy for the season
}
\description{
Calculates the Log Loss and Accuracy of the model by re-estimating m and rho daily and creating game odds
}
