% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getShortTeam}
\alias{getShortTeam}
\title{Get a Team's short code}
\usage{
getShortTeam(team)
}
\arguments{
\item{team}{a single team or vector of teams}
}
\value{
the team's (or teams') short code(s)
}
\description{
Get a Team's short code
}
