% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{getTeamColours}
\alias{getTeamColours}
\title{Get Team Colours}
\usage{
getTeamColours(
  home,
  away,
  delta = 0.15,
  teamColours = HockeyModel::teamColours
)
}
\arguments{
\item{home}{Home Team colours to get}

\item{away}{Away Team's colours to get}

\item{delta}{Colour delta required. Default 0.15. See \link{colourDelta}. Must be between 0 and 1}

\item{teamColours}{Team colours table. Defaults to \link{teamColours}. Pass
\link{pwhlTeamColours} for PWHL teams.}
}
\value{
a list with two items: home & away, each containing the appropriate hex colour value
}
\description{
Given a home and away team, return a set of two colours that correspond one to each team, yet are not too similar to eachother. Instead of returning Blue and Blue for Buffalo and Tampa, change Buffalo to their Gold
}
\examples{
getTeamColours("Buffalo Sabres", "Tampa Bay Lightning")
}
