% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interface.R
\name{getTeamDivisions}
\alias{getTeamDivisions}
\title{Get divisions for one or more teams}
\usage{
getTeamDivisions(teams, teamColours = HockeyModel::teamColours)
}
\arguments{
\item{teams}{(\code{character}) Team names.}

\item{teamColours}{(\code{data.frame}) Team metadata table.}
}
\value{
(\code{character}) Division name for each input team.
}
\description{
Get divisions for one or more teams
}
\keyword{internal}
