% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{historicalPoints}
\alias{historicalPoints}
\title{Get Historical Points for all teams listed in a scores frame}
\usage{
historicalPoints(sc)
}
\arguments{
\item{sc}{scores frame}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with season point totals by team.
}
\description{
Get Historical Points for all teams listed in a scores frame
}
