% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interface.R
\name{inOffSeason}
\alias{inOffSeason}
\title{In Off Season}
\usage{
inOffSeason(date = Sys.Date())
}
\arguments{
\item{date}{Date to check if it's in the (current) off season. Default today}
}
\value{
TRUE if we're in off-season, else FALSE
}
\description{
Determine if we're currently in the off-season
}
