% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interface.R
\name{inPlayoffs}
\alias{inPlayoffs}
\title{In Playoffs}
\usage{
inPlayoffs(date = Sys.Date(), boolean = TRUE)
}
\arguments{
\item{date}{Date to check if it's in a playoffs period}

\item{boolean}{Whether to return a result as TRUE/FALSE (True) or to return the seasonID if the date is in a playoffs}
}
\value{
Either TRUE/FALSE or a seasonID/FALSE
}
\description{
check if the date is in a playoff period. Note: playoffs are considered from the day after the regular
season ends.
}
