% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interface.R
\name{inRegularSeason}
\alias{inRegularSeason}
\title{inRegularSeason}
\usage{
inRegularSeason(date = Sys.Date())
}
\arguments{
\item{date}{Date to check if it's in the most recent season. Default today}
}
\value{
TRUE if currently in regular season, else FALSE
}
\description{
Determine if we're currently in Regular Season
}
