% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.Date}
\alias{is.Date}
\title{Is this a Date?}
\usage{
is.Date(date)
}
\arguments{
\item{date}{is this a date?}
}
\value{
is it a date?
}
\description{
Is this a date?
}
\examples{
is.Date("2020-12-13"); is.Date("bob")
}
