% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-api-fetch.R
\name{load_or_get_nst}
\alias{load_or_get_nst}
\title{Load or download a Natural Stat Trick game report}
\usage{
load_or_get_nst(gid)
}
\arguments{
\item{gid}{(\code{character(1)} or \code{numeric(1)}) NHL game ID in ten-digit format.}
}
\value{
(\code{data.frame}) Natural Stat Trick report rows for \code{gid}.
}
\description{
Load or download a Natural Stat Trick game report
}
\keyword{internal}
