% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{loopless_sim}
\alias{loopless_sim}
\title{'Loopless' simulation}
\usage{
loopless_sim(
  nsims = 1e+05,
  cores = parallel::detectCores() - 1,
  schedule = HockeyModel::schedule,
  scores = HockeyModel::scores,
  params = NULL,
  season_sofar = NULL,
  likelihood_graphic = TRUE
)
}
\arguments{
\item{nsims}{number of simulations to run (approximate)}

\item{cores}{number of cores in parallel to process}

\item{schedule}{games to play}

\item{scores}{Season to this point}

\item{params}{The named list containing m, rho, beta, eta, and k. See \link{updateDC} for information on the params list}

\item{season_sofar}{The results of the season to date}

\item{likelihood_graphic}{whether to create a likelihood graphic}
}
\value{
a two member list, of all results and summary results
}
\description{
'Loopless' simulation
}
