% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{loopless_sim}
\alias{loopless_sim}
\title{'Loopless' simulation}
\usage{
loopless_sim(nsims = 1e+05, cores = parallel::detectCores() - 1,
  odds_table = NULL, season_sofar = NULL)
}
\arguments{
\item{nsims}{number of simulations}

\item{cores}{number of cores in parallel to process}

\item{odds_table}{The odds table, if any}

\item{season_sofar}{Season to this point}
}
\value{
a two member list, of all results and summary results
}
\description{
'Loopless' simulation
}
