% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{loopless_sim}
\alias{loopless_sim}
\title{'Loopless' simulation}
\usage{
loopless_sim(
  nsims = 1e+05,
  cores = parallel::detectCores() - 1,
  schedule = HockeyModel::schedule,
  scores = HockeyModel::scores,
  rho = HockeyModel::rho,
  m = HockeyModel::m,
  odds_table = NULL,
  season_sofar = NULL
)
}
\arguments{
\item{nsims}{number of simulations to run (approximate)}

\item{cores}{number of cores in parallel to process}

\item{schedule}{games to play}

\item{scores}{Season to this point}

\item{rho}{rho factor to pass. default HockeyModel::rho}

\item{m}{m model to pass. default HockeyModel::m}

\item{odds_table}{odds from remainderSeasonDC(Odds = TRUE), or null}
}
\value{
a two member list, of all results and summary results
}
\description{
'Loopless' simulation
}
