% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mse}
\alias{mse}
\title{MSE}
\usage{
mse(predicted, actual)
}
\arguments{
\item{predicted}{Predicted numeric value}

\item{actual}{Actual numeric value}
}
\value{
a single value for MSE
}
\description{
calculate the MSE metrics. From MLMetrics
}
