% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhl_divisions}
\alias{nhl_divisions}
\title{NHL Divisions}
\format{
A list of vectors:
\describe{
  \item{Atlantic}{Teams in the Atlantic division}
  \item{Central}{Teams in the Central division}
  \item{Metropolitan}{Teams in the Metropolitan division}
  \item{Pacific}{Teams in the Pacific division}
}
}
\source{
\url{http://www.nhl.com/}
}
\usage{
nhl_divisions
}
\description{
A list of divisions with teams in each vector.
}
\keyword{datasets}
