% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_iterative.R
\name{optimizeIterative}
\alias{optimizeIterative}
\title{Optimize Iterative Dixon-Coles}
\usage{
optimizeIterative(
  target_model = "wl",
  params = HockeyModel::iterativeParameters
)
}
\arguments{
\item{target_model}{Default 'wl' for targeting optimization of the win/loss model. Also accepts 'xG'}

\item{params}{The starting point for Parameter optimization}
}
\value{
the results of \code{optim} including parameters
}
\description{
Call this to optimize the iterative dixon-coles method.
}
