% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseCores}
\alias{parseCores}
\title{Resolve a safe worker-core count}
\usage{
parseCores(cores)
}
\arguments{
\item{cores}{(\code{integer(1)} or \code{NULL}) Requested core count.}
}
\value{
(\code{integer(1)}) Core count constrained to available resources.
}
\description{
Resolve a safe worker-core count
}
\keyword{internal}
