% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{playoffDC}
\alias{playoffDC}
\title{Predict the odds of a playoff series}
\usage{
playoffDC(home, away, nsims = 1e+05, rho = HockeyModel::rho,
  m = HockeyModel::m)
}
\arguments{
\item{home}{The team with home ice advantage}

\item{away}{The opponent team}

\item{nsims}{number of simulations to run}

\item{rho}{HockeyModel::rho if nothing}

\item{m}{HockeyModel::m if nothing}
}
\value{
a list of home & away team odds for the playoff series
}
\description{
Predict the odds of a playoff series
}
