% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{playoffDC}
\alias{playoffDC}
\title{Playoff Odds DC}
\usage{
playoffDC(home, away, rho = HockeyModel::rho, m = HockeyModel::m,
  home_wins = 0, away_wins = 0)
}
\arguments{
\item{home}{Series Home Ice Advantage Team Name}

\item{away}{Away (Opponent) Team Name}

\item{rho}{HockeyModel::rho}

\item{m}{HockeyModel::m}

\item{home_wins}{Number of wins for home ice advantage team thus far in series}

\item{away_wins}{Number of wins for away team thus far in series}
}
\value{
home ice advantage team odds to win series
}
\description{
Playoff Odds DC
}
