% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{playoffOdds}
\alias{playoffOdds}
\title{Predict playoff odds graphic}
\usage{
playoffOdds()
}
\value{
A playoff-odds \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Convenience wrapper around \code{\link[=plot_prediction_playoffs_by_team]{plot_prediction_playoffs_by_team()}}.
}
