% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{playoffOdds}
\alias{playoffOdds}
\title{Predict playoff odds graphic}
\usage{
playoffOdds(...)
}
\arguments{
\item{...}{Items to pass to update and plot functions.}
}
\value{
playoff odds ggplot object
}
\description{
Predict playoff odds graphic
}
