% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{playoffSeriesOdds}
\alias{playoffSeriesOdds}
\title{Statistical Playoff Series Odds Solver}
\usage{
playoffSeriesOdds(home_odds, away_odds, home_win = 0, away_win = 0)
}
\arguments{
\item{home_odds}{Team odds with home-ice advantage at home}

\item{away_odds}{Team odds with home-ice advantage at away (on the road)}

\item{home_win}{Number of home ice advantage team wins thus far in the series. Default to 0 (prediction before series start)}

\item{away_win}{Number of away team wins thus far in the series}
}
\value{
numeric odds of home team win series (1-odds for away odds)
}
\description{
Given home and away win odds, produce the odds of the 'home advantage' team winning the series. From \url{http://www.stat.umn.edu/geyer/playoff.html}, modified to function with odds determination.
}
\references{
\url{http://www.stat.umn.edu/geyer/playoff.html}
}
