% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{playoffSolverEngine}
\alias{playoffSolverEngine}
\title{Playoff Solver Engine}
\usage{
playoffSolverEngine(
  nsims,
  completedSeries,
  east_results,
  west_results,
  currentSeries,
  summary_results,
  homeAwayOdds,
  params = NULL
)
}
\arguments{
\item{nsims}{number of sims (in each core)}

\item{completedSeries}{completed series}

\item{east_results}{east_results}

\item{west_results}{west_results}

\item{currentSeries}{currentSeries}

\item{summary_results}{summary_results}

\item{homeAwayOdds}{precalculated home & away pairs of odds - if available.}

\item{params}{The named list containing m, rho, beta, eta, and k. See \link{updateDC} for information on the params list}
}
\description{
Does the actual simulating. A function so it's parallelizable. Not to be called directly
}
