% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{playoffWin}
\alias{playoffWin}
\title{Playoff Win Calculator}
\usage{
playoffWin(
  home_team,
  away_team,
  home_wins = 0,
  away_wins = 0,
  m = HockeyModel::m,
  rho = HockeyModel::rho,
  theta = HockeyModel::theta,
  gamma = HockeyModel::gamma
)
}
\arguments{
\item{home_team}{Home Ice Advantage Team}

\item{away_team}{Opponent Team}

\item{home_wins}{Home Ice Advantage Team Wins in Series}

\item{away_wins}{Opponent Team Wins in Series}

\item{m}{HockeyModel::m}

\item{rho}{HockeyModel::rho}

\item{theta}{HockeyModel::theta}

\item{gamma}{HockeyModel::gamma}
}
\value{
Odds from 0-1 of home team winning. Away odds are 1 - return value
}
\description{
wraps PlayoffSeriesOdds with odds generator for a given home and away team
}
