% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics-plots.R
\name{plot_game}
\alias{plot_game}
\title{Plot single game expected goals}
\usage{
plot_game(home, away, params = NULL, maxgoal = 10)
}
\arguments{
\item{home}{The Home Team}

\item{away}{The Away Team}

\item{params}{The named list containing m, rho, beta, eta, and k. See \link{updateDC} for information on the params list}

\item{maxgoal}{the max number of goals to predict. Plot a few less.}
}
\value{
a ggplot object
}
\description{
Plot single game expected goals
}
