% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_game}
\alias{plot_game}
\title{Plot single game expected goals}
\usage{
plot_game(
  home,
  away,
  m = HockeyModel::m,
  rho = HockeyModel::rho,
  theta = HockeyModel::theta,
  gamma = HockeyModel::gamma,
  maxgoal = 10
)
}
\arguments{
\item{home}{The Home Team}

\item{away}{The Away Team}

\item{m}{the DC m model}

\item{rho}{the DC rho value}

\item{theta}{the DC theta value}

\item{gamma}{the DC gamma value}

\item{maxgoal}{the max number of goals to predict. Plot a few less.}
}
\value{
a ggplot object
}
\description{
Plot single game expected goals
}
