% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_game}
\alias{plot_game}
\title{Plot single game expected goals}
\usage{
plot_game(
  home,
  away,
  m = HockeyModel::m,
  rho = HockeyModel::rho,
  maxgoal = 10,
  teamColours = HockeyModel::teamColours
)
}
\arguments{
\item{home}{The Home Team}

\item{away}{The AWay Team}

\item{m}{the DC m model}

\item{rho}{the DC rho value}

\item{maxgoal}{the max number of goals to predict. Plot a few less.}

\item{teamColours}{HockeyModel::teamColours or a custom value}
}
\value{
a ggplot object
}
\description{
Plot single game expected goals
}
