% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{plot_game}
\alias{plot_game}
\title{Title}
\usage{
plot_game(home, away, m = HockeyModel::m, rho = HockeyModel::rho)
}
\arguments{
\item{home}{The Home Team}

\item{away}{The AWay Team}

\item{m}{the DC m model}

\item{rho}{the DC rho value}
}
\value{
a list of expected home & away goals, plus a plot.
}
\description{
Title
}
