% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_odds_today}
\alias{plot_odds_today}
\title{Plot Today's Odds}
\usage{
plot_odds_today(
  today = Sys.Date(),
  rho = HockeyModel::rho,
  m = HockeyModel::m,
  theta = HockeyModel::theta,
  gamma = HockeyModel::gamma,
  schedule = HockeyModel::schedule,
  teamColours = HockeyModel::teamColours,
  ...
)
}
\arguments{
\item{today}{The day's odds to plot. Default today.}

\item{rho}{HockeyModel::rho or a custom value}

\item{m}{HockeyModel::m or a custom value}

\item{theta}{HocekyModel::theta or a custom value}

\item{gamma}{HockeyModel::gamma or a custom value}

\item{schedule}{HockeyModel::schedule or a custom value}

\item{teamColours}{HockeyModel::teamColours or a custom value}

\item{...}{additional parameters to pass}
}
\value{
a ggplot image of odds
}
\description{
Plot Today's Odds
}
