% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{plot_odds_today}
\alias{plot_odds_today}
\title{Plot Today's Odds}
\usage{
plot_odds_today(today = Sys.Date(), rho = HockeyModel::rho,
  m = HockeyModel::m, schedule = HockeyModel::schedule, ...)
}
\arguments{
\item{today}{The day's odds to plot}

\item{...}{additional parameters to pass}
}
\value{
a ggplot image of odds
}
\description{
Plot Today's Odds
}
