% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{plot_odds_today}
\alias{plot_odds_today}
\title{Plot Today's Odds}
\usage{
plot_odds_today(today = Sys.Date(), schedule = HockeyModel::schedule,
  ...)
}
\arguments{
\item{today}{The day's odds to plot}

\item{schedule}{The schedule of games}

\item{...}{additional parameters to pass}
}
\description{
Plot Today's Odds
}
