% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{plot_pace_by_team}
\alias{plot_pace_by_team}
\title{Plot Pace by team}
\usage{
plot_pace_by_team(team = "all", scores = HockeyModel::scores,
  season = NULL)
}
\arguments{
\item{team}{Specific team(s) to plot. Default = all}

\item{scores}{the historical scores.}

\item{season}{choose a specific season. Default NULL is most recent (current)}
}
\value{
a ggplot object
}
\description{
Plot Pace by team
}
