% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_pace_by_team}
\alias{plot_pace_by_team}
\title{Plot Pace by team}
\usage{
plot_pace_by_team(
  graphic_dir = "./prediction_results/graphics",
  subdir = "pace",
  prediction_dir = "./prediction_results",
  scores = HockeyModel::scores,
  teamColours = HockeyModel::teamColours
)
}
\arguments{
\item{graphic_dir}{The graphics directory}

\item{subdir}{The pace subdirectory in graphics}

\item{prediction_dir}{The predictions directory}

\item{scores}{The HockeyModel::scores object, or custom scores in the same format.}

\item{teamColours}{HockeyModel::teamColours or a custom value}
}
\description{
Plot Pace by team
}
