% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_playoff_series_odds}
\alias{plot_playoff_series_odds}
\title{Plot Today's Playoff Series Odds}
\usage{
plot_playoff_series_odds(
  series = HockeyModel::series,
  rho = HockeyModel::rho,
  m = HockeyModel::m,
  theta = HockeyModel::theta,
  gamma = HockeyModel::gamma,
  schedule = HockeyModel::schedule,
  teamColours = HockeyModel::teamColours,
  ...
)
}
\arguments{
\item{series}{A data frame of home team, away team, home wins, away wins}

\item{rho}{HockeyModel::rho or a custom value}

\item{m}{HockeyModel::m or a custom value}

\item{theta}{HockekyModel::theta or a custom value}

\item{gamma}{HockeyModel::gamma or a custom value}

\item{schedule}{HockeyModel::schedule or a custom value}

\item{teamColours}{HockeyModel::teamColours or a custom value}

\item{...}{additional parameters to pass}
}
\value{
a ggplot image of odds
}
\description{
Plot Today's Playoff Series Odds
}
