% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_point_likelihood}
\alias{plot_point_likelihood}
\title{Team Point Predict Plot}
\usage{
plot_point_likelihood(
  preds = NULL,
  graphic_dir = "./prediction_results/graphics",
  subdir = "pace",
  ...
)
}
\arguments{
\item{preds}{Raw predictions to generate ggridges point likelyhood plot. Otherwise will generate with current m, rho, scores and schedule in HockeyModel.}

\item{graphic_dir}{Directory to save plot images}

\item{subdir}{Subdirectory to save plot images}

\item{...}{Additional parameters to pass to loopless sim}
}
\value{
two plots in list, as $eastplot and $westplot
}
\description{
Team Point Predict Plot
}
