% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_point_likelihood}
\alias{plot_point_likelihood}
\title{Team Point Predict Plot}
\usage{
plot_point_likelihood(
  preds = NULL,
  graphic_dir = file.path(devtools::package_file(), "prediction_results", "graphics"),
  subdir = "pace",
  savefiles = TRUE
)
}
\arguments{
\item{preds}{Raw predictions to generate ggridges point likelyhood plot.}

\item{graphic_dir}{Directory to save plot images}

\item{subdir}{Subdirectory to save plot images}

\item{savefiles}{Whether to save files to disk}
}
\value{
plot(s) in a list, named for conference(s) in use at the time.
}
\description{
Team Point Predict Plot
}
