% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{plot_prediction_points_by_team}
\alias{plot_prediction_points_by_team}
\title{Plot Predicted Points}
\usage{
plot_prediction_points_by_team(all_predictions = compile_predictions(),
  past_days = 14, teamColours = HockeyModel::teamColours)
}
\arguments{
\item{all_predictions}{the compiled predictions}

\item{past_days}{number of past days to include on the plot. Default: a fortnight}

\item{teamColours}{HockeyModel::teamColours or a custom value}
}
\value{
a ggplot object
}
\description{
Plot Predicted Points
}
