% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{plot_prediction_presidents_by_team}
\alias{plot_prediction_presidents_by_team}
\title{Plot President's Trophey Odds}
\usage{

  plot_prediction_presidents_by_team(all_predictions = compile_predictions(),
  past_days = 14)
}
\arguments{
\item{all_predictions}{the compiled predictions}

\item{past_days}{number of past days to include on the plot. Default: a fortnight}
}
\value{
a ggplot object
}
\description{
Plot President's Trophey Odds
}
