% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_team_rating}
\alias{plot_team_rating}
\title{Plot Team Rating}
\usage{
plot_team_rating(m = HockeyModel::m, teamlist = NULL)
}
\arguments{
\item{m}{HockeyModel::m}

\item{teamlist}{select a subset of teams, if desired}
}
\value{
a ggplot2 plot
}
\description{
Produces a plot of offensive and defensive ratings of teams, 0 centred (not scaled).
}
