% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pointPredict}
\alias{pointPredict}
\title{Predict Points graphic}
\usage{
pointPredict()
}
\value{
A point-projection \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Convenience wrapper around \code{\link[=plot_prediction_points_by_team]{plot_prediction_points_by_team()}}.
}
