% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pointPredict}
\alias{pointPredict}
\title{Predict Points graphic}
\usage{
pointPredict(...)
}
\arguments{
\item{...}{Items to pass to update and plot functions.}
}
\value{
point prediction ggplot object
}
\description{
Predict Points graphic
}
