% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley-terry.R
\name{prepareDataForBT}
\alias{prepareDataForBT}
\title{Prepare Data for BT fitting}
\usage{
prepareDataForBT(data = HockeyModel::scores)
}
\arguments{
\item{data}{The scores data to convert. Defaults to all HockeyModel::scores.}
}
\value{
a long data frame
}
\description{
Prepare Data for BT fitting
}
