% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{presidentOdds}
\alias{presidentOdds}
\title{Predict President's Odds graphic}
\usage{
presidentOdds()
}
\value{
A President's Trophy odds \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Convenience wrapper around \code{\link[=plot_prediction_presidents_by_team]{plot_prediction_presidents_by_team()}}.
}
