% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{presidentOdds}
\alias{presidentOdds}
\title{Predict President's Odds graphic}
\usage{
presidentOdds(...)
}
\arguments{
\item{...}{Items to pass to update and plot functions.}
}
\value{
president's odds ggplot object
}
\description{
Predict President's Odds graphic
}
