% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{prob_matrix}
\alias{prob_matrix}
\title{Probability Matrix}
\usage{
prob_matrix(lambda, mu, rho, theta, gamma, maxgoal)
}
\arguments{
\item{lambda}{home lambda}

\item{mu}{away mu}

\item{rho}{low score adjustment}

\item{theta}{tie game adjustment}

\item{gamma}{tie game multiplier}

\item{maxgoal}{max goals per game}
}
\value{
a square matrix of maxgoal:maxgoal
}
\description{
Given a mu, lambda, rho, and theta, generate a probability matrix. Differs from dcProbMatrix in that no regresson or solving for supplied teams happens
}
