% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhl-api-interface.R
\name{pwhl_api_request}
\alias{pwhl_api_request}
\title{Build a PWHL HockeyTech API request}
\usage{
pwhl_api_request(params)
}
\arguments{
\item{params}{Named list of query parameters to append.}
}
\value{
A prepared \code{\link[httr2:request]{httr2::request}} object.
}
\description{
Build a PWHL HockeyTech API request
}
\keyword{internal}
