% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhl-api-interface.R
\name{pwhl_games_today}
\alias{pwhl_games_today}
\title{Get today's PWHL games}
\usage{
pwhl_games_today(schedule = HockeyModel::pwhlSchedule, date = Sys.Date())
}
\arguments{
\item{schedule}{(\code{data.frame}) PWHL schedule, defaults to
\link{pwhlSchedule}.}

\item{date}{(\code{Date}) The date to look up. Defaults to today.}
}
\value{
A data frame of scheduled PWHL games for that date, or \code{NULL}.
}
\description{
Returns PWHL games scheduled for a given date, or \code{NULL} if
there are none.
}
