% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwhl_k}
\alias{pwhl_k}
\title{PWHL \code{k}}
\format{
a single numeric value, or \code{NULL} when not yet fitted
}
\usage{
pwhl_k
}
\description{
Weibull multiplier for PWHL tie-game enhancement. Populated by
\code{\link[=updatePWHLDC]{updatePWHLDC()}}.
}
\keyword{datasets}
