% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhl-model.R
\name{pwhl_season_start_date}
\alias{pwhl_season_start_date}
\title{PWHL season start date}
\usage{
pwhl_season_start_date(schedule = HockeyModel::pwhlSchedule)
}
\arguments{
\item{schedule}{(\code{data.frame}) PWHL schedule. Defaults to
\link{pwhlSchedule}.}
}
\value{
(\code{Date}) First game date in the schedule, or \code{NA} if empty.
}
\description{
PWHL season start date
}
\keyword{internal}
