% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{randomSeriesWinner}
\alias{randomSeriesWinner}
\title{Random Series Winner}
\usage{
randomSeriesWinner(
  home_team,
  away_team,
  home_wins = 0,
  away_wins = 0,
  homeAwayOdds = NULL,
  params = NULL
)
}
\arguments{
\item{home_team}{Home Team name (required)}

\item{away_team}{Away Team name (Required)}

\item{home_wins}{Number of home wins (default 0)}

\item{away_wins}{Number of away team wins (default 0)}

\item{homeAwayOdds}{pre-calculated home & away team parings odds of a home win. Overrides playoffwin calculation}

\item{params}{The named list containing m, rho, beta, eta, and k. See \link{updateDC} for information on the params list}
}
\value{
TRUE if the home team wins, else FALSE
}
\description{
generate a random series winner given a home and away team
}
