% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ratings}
\alias{ratings}
\title{Current ratings}
\usage{
ratings(m = HockeyModel::m, ...)
}
\arguments{
\item{m}{HockeyModel::m or a custom value}

\item{...}{Items to pass to plot functions.}
}
\value{
today's ratings ggplot object
}
\description{
Current ratings
}
