% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-league-predictions-log.R
\name{recordTodaysPredictions}
\alias{recordTodaysPredictions}
\title{Record Today's Predictions}
\usage{
recordTodaysPredictions(
  today = Sys.Date(),
  filepath = file.path(getOption("HockeyModel.data.path"), "dailyodds.csv"),
  schedule = HockeyModel::schedule,
  params = NULL,
  include_xG = FALSE,
  draws = TRUE
)
}
\arguments{
\item{today}{Day's predictions to record. Defaults to today, but can set any other day}

\item{filepath}{csv file location to store predictions. Will append to file.}

\item{schedule}{HockeyModel::schedule or supplied. \code{today} date must be in schedule}

\item{params}{The named list containing m, rho, beta, eta, and k. See \link{updateDC} for information on the params list}

\item{include_xG}{Whether to record daily XG values}

\item{draws}{whether to record draw odds (True) or not (False). Default is True}
}
\description{
Record today's predictions to file (for easy later retrieval). Run \code{cleanupPredictionsFile} periodically to tidy
}
