% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{recovery}
\alias{recovery}
\title{Recovery Calculating}
\usage{
recovery(pPoints, sdPoints, aPoints)
}
\arguments{
\item{predicted}{data.frame of Team, Predicted Points, SD}

\item{actual}{data.frame of Team, Actual Points}
}
\value{
single value 0-31
}
\description{
Recovery Calculating
}
