% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley-terry (original).R, R/bradley-terry.R
\name{remainderSeasonBT}
\alias{remainderSeasonBT}
\title{BT remainder of season}
\usage{
remainderSeasonBT(nsims = 10000, scores = HockeyModel::scores,
  schedule = HockeyModel::schedule)

remainderSeasonBT(nsims = 10000, scores = HockeyModel::scores,
  schedule = HockeyModel::schedule)
}
\arguments{
\item{nsims}{Number of simulations}

\item{scores}{the historical scores}

\item{schedule}{uplayed future games}

\item{nsims}{Number of simulations}

\item{scores}{the historical scores}

\item{schedule}{uplayed future games}
}
\value{
data frame of Team, playoff odds.

data frame of Team, playoff odds.
}
\description{
Odds for each team to get to playoffs.

Odds for each team to get to playoffs.
}
