% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{remainderSeasonDC}
\alias{remainderSeasonDC}
\title{DC remainder of season}
\usage{
remainderSeasonDC(nsims = 10, scores = HockeyModel::scores,
  schedule = HockeyModel::schedule)
}
\arguments{
\item{nsims}{Number of simulations}

\item{scores}{the historical scores}

\item{schedule}{uplayed future games}
}
\value{
data frame of Team, playoff odds.
}
\description{
Odds for each team to get to playoffs.
}
