% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{remainderSeasonDC}
\alias{remainderSeasonDC}
\title{DC remainder of season}
\usage{
remainderSeasonDC(nsims = 10000, scores = HockeyModel::scores,
  schedule = HockeyModel::schedule, odds = FALSE, regress = FALSE,
  ...)
}
\arguments{
\item{nsims}{Number of simulations}

\item{scores}{the historical scores}

\item{schedule}{uplayed future games}

\item{odds}{whether to return odds table or simulate season}

\item{regress}{whether to apply a regression to the mean for team strength on future predictions}

\item{...}{arguements to pass to dc predictor}
}
\value{
data frame of Team, playoff odds.
}
\description{
Odds for each team to get to playoffs.
}
