% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scores}
\alias{scores}
\title{Historical NHL and WHA scores.}
\format{
A data frame with  variables:
\describe{
\item{Date}{Date of the game}
\item{AwayTeam}{Name of visiting team}
\item{AwayGoals}{Number of goals for the visiting team}
\item{HomeTeam}{Name of the home team}
\item{HomeGoals}{Number of goals for the home team}
\item{OTStatus}{The OT status of the game (none, OT, SO, 2OT, 3OT, ...)}
\item{GameID}{Unique GameID for each game}
\item{GameType}{Game Type for each game}
\item{GameStatus}{Game Status - 'Final'}
\item{Result}{A numerically coded result for Hoome Team. Win=1, OTWin = 0.75, SOWin = 0.6, Tie = 0.5, SOLoss = 0.4, OTLoss = 0.25, Loss = 0}
\item{HomexG}{Home Team's xG}
\item{AwayxG}{Away Team's xG}
\item{AwayCF}{Advanced Stats Properties}
\item{AwayCFpk}{Advanced Stats Properties}
\item{AwayCFpp}{Advanced Stats Properties}
\item{AwayG}{Advanced Stats Properties}
\item{AwayGpk}{Advanced Stats Properties}
\item{AwayGpp}{Advanced Stats Properties}
\item{AwayxG}{Advanced Stats Properties}
\item{AwayxGpk}{Advanced Stats Properties}
\item{AwayxGpp}{Advanced Stats Properties}
\item{HomeCF}{Advanced Stats Properties}
\item{HomeCFpk}{Advanced Stats Properties}
\item{HomeCFpp}{Advanced Stats Properties}
\item{HomeG}{Advanced Stats Properties}
\item{HomeGpk}{Advanced Stats Properties}
\item{HomeGpp}{Advanced Stats Properties}
\item{HomexGpk}{Advanced Stats Properties}
\item{HomexGpp}{Advanced Stats Properties}
}
}
\source{
\url{http://www.hockey-reference.com/}
}
\usage{
scores
}
\description{
A dataset containing historical NHL and WHA scores.
}
\keyword{datasets}
