% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scores}
\alias{scores}
\title{Historical NHL and WHA scores.}
\format{A data frame with  variables:
\describe{
  \item{Date}{Date of the game}
  \item{Visitor}{Name of visiting team}
  \item{VisitorGoals}{Number of goals for the visiting team}
  \item{Home}{Name of the home team}
  \item{HomeGoals}{Number of goals for the home team}
  \item{OTStatus}{The OT status of the game (none, OT, SO, 2OT, 3OT, ...)}
  \item{League}{NHL or WHA}
  \item{Tie}{Boolean, Tie or Not}
  \item{Winner}{Name of the winning team}
  \item{Loser}{Name of the losing team}
  \item{Result}{Coded home win = 0, visitor win = 1}
}}
\source{
\url{http://www.hockey-reference.com/}
}
\usage{
scores
}
\description{
A dataset containing historical NHL and WHA scores.
}
\keyword{datasets}
