% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{series_odds_table}
\alias{series_odds_table}
\title{Series Odds Table}
\usage{
series_odds_table(series = getAPISeries(), params = NULL, include_logo = FALSE)
}
\arguments{
\item{series}{A data frame with columns HomeTeam, AwayTeam, HomeWins, and
AwayWins describing the current playoff series. Defaults to
\code{\link[=getAPISeries]{getAPISeries()}}.}

\item{params}{The named list containing m, rho, beta, eta, and k. See
\link{updateDC} for information on the params list.}

\item{include_logo}{Whether to include the Daily Faceoff logo as a source
note. Default \code{FALSE}.}
}
\value{
a gt table
}
\description{
Returns a gt table of odds for each active playoff series.
}
