% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{sim_engine}
\alias{sim_engine}
\title{Simulation engine to be parallelized or used in single core}
\usage{
sim_engine(all_season, nsims)
}
\arguments{
\item{all_season}{One seasons' scores & odds schedule}

\item{nsims}{Number of simulations to run}
}
\value{
results of each season's simulation, as one long data frame score table.
}
\description{
Simulation engine to be parallelized or used in single core
}
