% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley-terry.R
\name{simpleBTPredict}
\alias{simpleBTPredict}
\title{Simple BT Predict Game}
\usage{
simpleBTPredict(btModel, home.team, away.team)
}
\arguments{
\item{home.team}{text home team}

\item{away.team}{text away team}

\item{btSimple}{simple Bradley Terry Model}
}
\value{
odd for a home team win from 0-1
}
\description{
Simple BT Predict Game
}
