% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{simulatePlayoffs}
\alias{simulatePlayoffs}
\title{simulate Playoffs}
\usage{
simulatePlayoffs(
  summary_results = NULL,
  nsims = 1e+05,
  cores = parallel::detectCores() - 1,
  m = HockeyModel::m,
  rho = HockeyModel::rho
)
}
\arguments{
\item{summary_results}{summary results}

\item{nsims}{Number of playoff sims to run. Too many takes a long time.}

\item{cores}{Number of processor cores to use}

\item{m}{HockeyModel::m}

\item{rho}{HockeyModel::rho}
}
\value{
a data frame of each teams' odds of winning each round (First Round, Second Round, Conference Finals and Stanley Cup)
}
\description{
Solves playoff odds by MC simulation.
}
