% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-league-simulation.R
\name{simulateSeasonParallel}
\alias{simulateSeasonParallel}
\title{Simulate the remainder of the season}
\usage{
simulateSeasonParallel(
  scores = HockeyModel::scores,
  nsims = 10000,
  schedule = HockeyModel::schedule,
  cores = NULL,
  progress = FALSE,
  params = NULL
)
}
\arguments{
\item{scores}{Past (historical) season scores. Defaults to HockeyModel::Scores}

\item{nsims}{number of simulations to run}

\item{schedule}{Future unplayed games. Defaults to HockeyModel::schedule}

\item{cores}{number of cores to use in parallel.}

\item{progress}{whether to show a progress bar.}

\item{params}{The named list containing m, rho, beta, eta, and k. See \link{updateDC} for information on the params list}
}
\value{
a data frame of results
}
\description{
Simulate the remainder of the season
}
