% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl-league-playoffs.R
\name{single_series_solver}
\alias{single_series_solver}
\title{Given current seriess and a series number and home and away teams, either return true series winner, random series winner (with home/away wins considered) or random series winner}
\usage{
single_series_solver(
  series_number,
  currentSeries,
  homeTeam,
  awayTeam,
  homeAwayOdds = NULL,
  params = NULL
)
}
\arguments{
\item{series_number}{Series number from 1:15}

\item{currentSeries}{the full list of series returned from getAPISeries}

\item{homeTeam}{Home Team extracted from summary_results}

\item{awayTeam}{away Team extracted from summary_results}

\item{homeAwayOdds}{if calculated, the odds of a home or away team win}

\item{params}{The named list containing m, rho, beta, eta, and k. See \link{updateDC} for information on the params list}
}
\value{
a series winner (team name)
}
\description{
Given current seriess and a series number and home and away teams, either return true series winner, random series winner (with home/away wins considered) or random series winner
}
