% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{tau_singular}
\alias{tau_singular}
\title{Compute one Dixon-Coles low-score adjustment factor}
\usage{
tau_singular(xx, yy, lambda, mu, rho)
}
\arguments{
\item{xx}{(\code{integer(1)}) Home goals.}

\item{yy}{(\code{integer(1)}) Away goals.}

\item{lambda}{(\code{double(1)}) Home expected goals.}

\item{mu}{(\code{double(1)}) Away expected goals.}

\item{rho}{(\code{double(1)}) Low-score dependence parameter.}
}
\value{
(\code{double(1)}) Tau scaling factor for one score pair.
}
\description{
Compute one Dixon-Coles low-score adjustment factor
}
\keyword{internal}
