% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley-terry.R
\name{todayBT}
\alias{todayBT}
\title{Predict Today's results using BT}
\usage{
todayBT(today = Sys.Date(), fittedBT = HockeyModel::fittedBT,
  schedule = HockeyModel::schedule)
}
\arguments{
\item{today}{Date to predict, optional}

\item{fittedBT}{The bt model (Generalized Davidson or simple BT) as HockeyModel::fittedBT or HockeyModel::fittedBTSimple}

\item{schedule}{the future's schedule from HockeyModel::schedule}
}
\value{
a data frame of home win, away win, and draw odds.
}
\description{
Predict Today's results using BT
}
