% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{todayDC}
\alias{todayDC}
\title{DC Predictions Today}
\usage{
todayDC(
  today = Sys.Date(),
  rho = HockeyModel::rho,
  m = HockeyModel::m,
  schedule = HockeyModel::schedule,
  expected_mean = NULL,
  season_percent = NULL
)
}
\arguments{
\item{today}{Generate predictions for this date. Defaults to today}

\item{rho}{DC Rho}

\item{m}{DC m}

\item{schedule}{shcedule to use, if not the builtin}

\item{expected_mean}{the mean lambda & mu, used only for regression}

\item{season_percent}{the percent complete of the season, used for regression}
}
\value{
a data frame of HomeTeam, AwayTeam, HomeWin, AwayWin, Draw, or NULL if no games today
}
\description{
DC Predictions Today
}
