% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{todayDC}
\alias{todayDC}
\title{DC Predictions Today}
\usage{
todayDC(today = Sys.Date(), rho = HockeyModel::rho,
  m = HockeyModel::m, schedule = HockeyModel::schedule)
}
\arguments{
\item{today}{Generate predictions for this date. Defaults to today}

\item{rho}{DC Rho}

\item{m}{DC m}

\item{schedule}{shcedule to use, if not the builtin}
}
\value{
a data frame of HomeTeam, AwayTeam, HomeWin, AwayWin, Draw, or NULL if no games today
}
\description{
DC Predictions Today
}
