% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{todayDC}
\alias{todayDC}
\title{DC Predictions Today}
\usage{
todayDC(today = Sys.Date())
}
\arguments{
\item{today}{Generate predictions for this date. Defaults to today}
}
\value{
a data frame of HomeTeam, AwayTeam, HomeWin, AwayWin, Draw, or NULL if no games today
}
\description{
DC Predictions Today
}
