% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{todayOdds}
\alias{todayOdds}
\title{Today's game odds graphic}
\usage{
todayOdds(date = Sys.Date(), rho = HockeyModel::rho,
  m = HockeyModel::m, schedule = HockeyModel::schedule,
  scores = HockeyModel::scores, ...)
}
\arguments{
\item{date}{date to predict odds. Default today}

\item{rho}{HockeyModel::rho or a custom value}

\item{m}{HockeyModel::m or a custom value}

\item{schedule}{HockeyModel::schedule or a custom value}

\item{scores}{HockeyModel::scores or a custom value}

\item{...}{Items to pass to update functions.}
}
\value{
today's odds ggplot object
}
\description{
Today's game odds graphic
}
