% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{todayOdds}
\alias{todayOdds}
\title{Today's game odds graphic}
\usage{
todayOdds(date = Sys.Date(), ...)
}
\arguments{
\item{date}{date to predict odds. Default today}

\item{...}{Items to pass to update functions.}
}
\value{
today's odds ggplot object
}
\description{
Today's game odds graphic
}
