% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{todayOddsPlot}
\alias{todayOddsPlot}
\title{Today's game odds graphic}
\usage{
todayOddsPlot(
  date = Sys.Date(),
  params = NULL,
  schedule = HockeyModel::schedule,
  scores = HockeyModel::scores
)
}
\arguments{
\item{date}{date to predict odds. Default today}

\item{params}{The named list containing m, rho, beta, eta, and k. See \link{updateDC} for information on the params list}

\item{schedule}{HockeyModel::schedule or a custom value}

\item{scores}{HockeyModel::scores or a custom value}
}
\value{
today's odds ggplot object
}
\description{
Today's game odds graphic
}
