% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_weights.R
\name{tune_dc_weight}
\alias{tune_dc_weight}
\title{Evaluate a Dixon-Coles weighting configuration}
\usage{
tune_dc_weight(xi = 0.003, upsilon = 150)
}
\arguments{
\item{xi}{(\code{double(1)}) Time-decay slope for historical weighting.}

\item{upsilon}{(\code{double(1)}) Midpoint for the logistic weighting curve.}
}
\value{
(\code{data.frame}) Evaluation schedule with model probabilities for the
tested period.
}
\description{
Evaluate a Dixon-Coles weighting configuration
}
\keyword{internal}
