% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tweet}
\alias{tweet}
\title{Post daily model graphics to social media}
\usage{
tweet(
  games,
  graphic_dir = "./prediction_results/graphics",
  delay = stats::runif(1, min = 2, max = 6) * 60,
  schedule = HockeyModel::schedule
)
}
\arguments{
\item{games}{(\code{data.frame}) Games to reference for game-specific posts.}

\item{graphic_dir}{(\code{character(1)}) Directory containing generated images.}

\item{delay}{(\code{double(1)}) Delay in seconds between posts.}

\item{schedule}{(\code{data.frame}) Schedule used to determine daily context.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Post daily model graphics to social media
}
\keyword{internal}
