% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-social.R
\name{tweetGames}
\alias{tweetGames}
\title{Tweet Game Plots}
\usage{
tweetGames(
  games = games_today(),
  delay = stats::runif(1, min = 4, max = 8) * 60,
  graphic_dir = getOption("HockeyModel.graphics.path"),
  params = NULL
)
}
\arguments{
\item{games}{Games to tweet graphics from}

\item{delay}{Delay between tweets}

\item{graphic_dir}{the graphics directory}

\item{params}{The named list containing m, rho, beta, eta, and k. See \link{updateDC} for information on the params list}
}
\description{
Tweet Game Plots
}
