% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tweetGames}
\alias{tweetGames}
\title{Tweet Game Plots}
\usage{
tweetGames(
  games = games_today(),
  delay = stats::runif(1, min = 4, max = 8) * 60,
  graphic_dir = "./prediction_results/graphics/",
  m = HockeyModel::m,
  rho = HockeyModel::rho,
  theta = HockeyModel::theta,
  gamma = HockeyModel::gamma,
  token = rtweet::get_token()
)
}
\arguments{
\item{games}{Games to tweet graphics from}

\item{delay}{Delay between tweets}

\item{graphic_dir}{the graphics directory}

\item{m}{HockeyModel::m or a custom value}

\item{rho}{HockeyModel::rho or a custom value}

\item{theta}{HockeyModel::theta or a custom value}

\item{gamma}{HockeyModel::gamma or a custom value}

\item{token}{the token for rtweet}
}
\description{
Tweet Game Plots
}
