% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tweetGames}
\alias{tweetGames}
\title{Tweet Game Plots}
\usage{
tweetGames(games = HockeyModel::schedule[HockeyModel::schedule$Date ==
  Sys.Date(), ], delay = 60 * 15,
  graphic_dir = "./prediction_results/graphics/", m = HockeyModel::m,
  rho = HockeyModel::rho, token = rtweet::get_token(),
  teamColours = HockeyModel::teamColours)
}
\arguments{
\item{games}{Games to tweet graphics from}

\item{delay}{Delay between tweets}

\item{graphic_dir}{the graphics directory}

\item{m}{HockeyModel::m or a custom value}

\item{rho}{HockeyModel::rho or a custom value}

\item{token}{the token for rtweet}

\item{teamColours}{HockeyModel::teamColours or a custom value}
}
\description{
Tweet Game Plots
}
