% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tweetGames}
\alias{tweetGames}
\title{Tweet Game Plots}
\usage{
tweetGames(
  games = games_today(),
  delay = stats::runif(1, min = 4, max = 8) * 60,
  graphic_dir = file.path(devtools::package_file(), "prediction_results", "graphics"),
  params = NULL,
  token = rtweet::auth_get()
)
}
\arguments{
\item{games}{Games to tweet graphics from}

\item{delay}{Delay between tweets}

\item{graphic_dir}{the graphics directory}

\item{params}{The named list containing m, rho, beta, eta, and k. See \link{updateDC} for information on the params list}

\item{token}{the token for rtweet}
}
\description{
Tweet Game Plots
}
