% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-social.R
\name{tweetLikelihoods}
\alias{tweetLikelihoods}
\title{Tweet Likelihood plots (ggridges)}
\usage{
tweetLikelihoods(
  delay = stats::runif(1, min = 3, max = 6) * 60,
  graphic_dir = getOption("HockeyModel.graphics.path"),
  subdir = "pace",
  scores = HockeyModel::scores
)
}
\arguments{
\item{delay}{time to delay. Default 5 min}

\item{graphic_dir}{graphics directory}

\item{subdir}{subdirectory - usually 'preds'}

\item{scores}{updated scores}
}
\value{
(\code{data.frame}) A summary of each attempted post (see
\code{.summarize_post_results()}), invisibly.
}
\description{
Tweet Likelihood plots (ggridges)
}
