% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tweetLikelihoods}
\alias{tweetLikelihoods}
\title{Tweet Likelihood plots (ggridges)}
\usage{
tweetLikelihoods(delay = 60 * 5,
  graphic_dir = "./prediction_results/graphics/", subdir = "pace",
  token = rtweet::get_token(), scores = HockeyModel::scores)
}
\arguments{
\item{delay}{time to delay. Default 5 min}

\item{graphic_dir}{graphics directory}

\item{subdir}{subdirectory - usually 'preds'}

\item{token}{rtweeet token}

\item{scores}{updated scores}
}
\description{
Tweet Likelihood plots (ggridges)
}
