% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tweetLikelihoods}
\alias{tweetLikelihoods}
\title{Tweet Likelihood plots (ggridges)}
\usage{
tweetLikelihoods(
  delay = stats::runif(1, min = 3, max = 6) * 60,
  graphic_dir = getOption("HockeyModel.graphics.path"),
  subdir = "pace",
  twitter_token = NULL,
  scores = HockeyModel::scores
)
}
\arguments{
\item{delay}{time to delay. Default 5 min}

\item{graphic_dir}{graphics directory}

\item{subdir}{subdirectory - usually 'preds'}

\item{twitter_token}{rtweeet token}

\item{scores}{updated scores}
}
\description{
Tweet Likelihood plots (ggridges)
}
