% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tweetMetrics}
\alias{tweetMetrics}
\title{Tweet Metrics}
\usage{
tweetMetrics(twitter_token = NULL)
}
\arguments{
\item{twitter_token}{rtweet token}
}
\description{
Tweet the metrics (Log Loss and Accuracy)
}
