% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tweetPace}
\alias{tweetPace}
\title{Tweet Pace Plots}
\usage{
tweetPace(delay = 60 * 5,
  graphic_dir = "./prediction_results/graphics/", subdir = "pace",
  prediction_dir = "./prediction_results/",
  token = rtweet::get_token(), scores = HockeyModel::scores)
}
\arguments{
\item{delay}{Delay between posted tweets}

\item{graphic_dir}{The graphics directory}

\item{subdir}{The pace subdirectory in graphics}

\item{prediction_dir}{The predictions directory}
}
\description{
Tweet Pace Plots
}
