% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-social.R
\name{tweetPlayoffOdds}
\alias{tweetPlayoffOdds}
\title{Tweet Playoff Odds}
\usage{
tweetPlayoffOdds(
  summary_results = NULL,
  params = NULL,
  graphic_dir = getOption("HockeyModel.graphics.path"),
  trimcup = FALSE
)
}
\arguments{
\item{summary_results}{the summary results file, otherwise the most recent will be loaded}

\item{params}{The named list containing m, rho, beta, eta, and k. See \link{updateDC} for information on the params list}

\item{graphic_dir}{graphic dir}

\item{trimcup}{trim to just cup winners}
}
\description{
Tweet a graphic of the playoff odds
}
