% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tweetPlayoffOdds}
\alias{tweetPlayoffOdds}
\title{Tweet Playoff Odds}
\usage{
tweetPlayoffOdds(
  summary_results = NULL,
  m = HockeyModel::m,
  rho = HockeyModel::rho,
  theta = HockeyModel::theta,
  gamma = HockeyModel::gamma,
  token = rtweet::get_token(),
  graphic_dir = "./prediction_results/graphics/",
  trimcup = FALSE
)
}
\arguments{
\item{summary_results}{the summary results file, otherwise the msot recent will be loaded}

\item{m}{HockeyModel::m or supplied}

\item{rho}{HockeyModel::rho or supplied}

\item{theta}{HockeyModel::theta or supplied}

\item{gamma}{HockeyModel::gamma or supplied}

\item{token}{token for twitter}

\item{graphic_dir}{graphic dir}

\item{trimcup}{trim to just cup winners}
}
\description{
Tweet a graphic of the playoff odds
}
