% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tweetPlayoffOdds}
\alias{tweetPlayoffOdds}
\title{Tweet Playoff Odds}
\usage{
tweetPlayoffOdds(
  playoffOdds = covid_play_in_solver(),
  token = rtweet::get_token(),
  graphic_dir = "./prediction_results/graphics/"
)
}
\arguments{
\item{playoffOdds}{During COVID, pass covid_play_in_solver(), afterwards change to all_results = all_results}

\item{token}{rtweet token}

\item{graphic_dir}{directory to save images}
}
\description{
Tweet the formattable graphics with each team's odds of each round/cup win
}
