% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tweetPlayoffOdds}
\alias{tweetPlayoffOdds}
\title{Tweet Playoff Odds}
\usage{
tweetPlayoffOdds(
  all_results = NULL,
  token = rtweet::get_token(),
  graphic_dir = "./prediction_results/graphics/"
)
}
\arguments{
\item{all_results}{all results, if not most recent (will be loaded by functions)}

\item{token}{rtweet token}

\item{graphic_dir}{directory to save images}
}
\description{
Tweet the \code{gt} graphics with each team's odds of each round/cup win
}
