% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tweetSeries}
\alias{tweetSeries}
\title{Tweet Series}
\usage{
tweetSeries(
  token = rtweet::get_token(),
  m = HockeyModel::m,
  rho = HockeyModel::rho,
  theta = HockeyModel::theta,
  gamma = HockeyModel::gamma,
  graphic_dir = "./prediction_results/graphics/"
)
}
\arguments{
\item{token}{rtweet token}

\item{m}{HockeyModel::m or supplied}

\item{rho}{HockeyModel::rho or supplied}

\item{theta}{HockeyModel::theta or supplied}

\item{gamma}{HockeyModel::gamma or supplied}

\item{graphic_dir}{directory to save the image}
}
\description{
Tweet the series odds graphics
}
