% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tweetSeries}
\alias{tweetSeries}
\title{Tweet Series}
\usage{
tweetSeries(
  token = rtweet::get_token(),
  params = NULL,
  graphic_dir = "./prediction_results/graphics/"
)
}
\arguments{
\item{token}{rtweet token}

\item{params}{The named list containing m, rho, beta, eta, and k. See \link{updateDC} for information on the params list}

\item{graphic_dir}{directory to save the image}
}
\description{
Tweet the series odds graphics
}
