% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tweetSeries}
\alias{tweetSeries}
\title{Tweet Series}
\usage{
tweetSeries(
  series = HockeyModel::series,
  token = rtweet::get_token(),
  graphic_dir = "./prediction_results/graphics/"
)
}
\arguments{
\item{series}{series data}

\item{token}{rtweet token}

\item{graphic_dir}{directory to save the image}
}
\description{
Tweet the series odds graphics
}
