% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon-coles.R
\name{updateDC}
\alias{updateDC}
\title{Update Dixon Coles parameters}
\usage{
updateDC(scores = HockeyModel::scores)
}
\arguments{
\item{scores}{scores, if not then data(scores)}
}
\description{
Update Dixon Coles parameters
}
