% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{updateModel}
\alias{updateModel}
\title{Update Model}
\usage{
updateModel(save_data = TRUE)
}
\arguments{
\item{save_data}{whether to save data to the package file}
}
\value{
a list of scores, schedule, and param named list
}
\description{
updates the schedule, scores, and model parameters (m, rho, theta, gamma). Returns them in a list of named values
}
