% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{updateModel}
\alias{updateModel}
\title{Update Model}
\usage{
updateModel()
}
\value{
a list of scores, schedule, m, rho, theta and gamma values. Values may be data frames, model output, or numeric
}
\description{
updates the schedule, scores, and model parameters (m, rho, theta, gamma). Returns them in a list of named values
}
