% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{updatePredictions}
\alias{updatePredictions}
\title{Update predictions}
\usage{
updatePredictions(data_dir = "./prediction_results/", ...)
}
\arguments{
\item{data_dir}{directory of predictions}

\item{...}{Items to pass to update functions.}
}
\description{
Update predictions
}
