% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{updatePredictions}
\alias{updatePredictions}
\title{Update predictions}
\usage{
updatePredictions(
  data_dir = "./prediction_results/",
  scores = HockeyModel::scores,
  schedule = HockeyModel::schedule,
  m = HockeyModel::m,
  rho = HockeyModel::rho,
  theta = HockeyModel::theta,
  gamma = HockeyModel::gamma
)
}
\arguments{
\item{data_dir}{directory of predictions}

\item{scores}{HockeyModel::scores or a custom value}

\item{schedule}{HockeyModel::schedule or a custom value}

\item{m}{HockeyModel::m or supplied}

\item{rho}{HockeyModel::rho or supplied}

\item{theta}{HockeyModel::theta or supplied}

\item{gamma}{HockeyModel::gamma or supplied}
}
\description{
Update predictions
}
