% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{updatePredictions}
\alias{updatePredictions}
\title{Update predictions}
\usage{
updatePredictions(data_dir = "./prediction_results/",
  scores = HockeyModel::scores, schedule = HockeyModel::schedule, ...)
}
\arguments{
\item{data_dir}{directory of predictions}

\item{scores}{HockeyModel::scores or a custom value}

\item{schedule}{HockeyModel::schedule or a custom value}

\item{...}{Items to pass to update functions.}
}
\description{
Update predictions
}
