% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{updateSchedule}
\alias{updateSchedule}
\title{Update Schedule}
\usage{
updateSchedule(data_dir = "./data-raw")
}
\arguments{
\item{data_dir}{The data storage directory}
}
\value{
True, if successful update or validation, `schedule` is a built in data
}
\description{
Updates the stored schedule if any games are rescheduled due to any event. Adds playoff games as the schedules are released
}
