% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interface.R
\name{updateScheduleAPI}
\alias{updateScheduleAPI}
\title{Update schedule using the NHL API}
\usage{
updateScheduleAPI(save_data = FALSE)
}
\arguments{
\item{save_data}{whether to write to package data}
}
\value{
data frame of schedule, after optionally writing to package data
}
\description{
Update schedule using the NHL API
}
