% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{updateScores}
\alias{updateScores}
\title{Update Scores}
\usage{
updateScores(data_dir = "./data-raw")
}
\arguments{
\item{data_dir}{The data storage directory}
}
\value{
True, if successful update or validation, `scores` is a built in data
}
\description{
Updates the scores of all finished games.
}
