% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interface.R
\name{updateScoresAPI}
\alias{updateScoresAPI}
\title{Update past scores using the NHL API}
\usage{
updateScoresAPI(
  scores = HockeyModel::scores,
  schedule = HockeyModel::schedule,
  full_season = FALSE,
  save_data = FALSE
)
}
\arguments{
\item{scores}{old scores}

\item{schedule}{current schedule}

\item{full_season}{whether to re-scrape the full season}

\item{save_data}{whether to write the data to package}
}
\value{
data frame of scores, after optionally writing to package data
}
\description{
Update past scores using the NHL API
}
